### 2021-06-08 #3343 @whedon
[PRE REVIEW]: Kinetics Toolkit: An Open-Source Python Package to Facilitate Research in Biomechanics
**Submitting author:** @felixchenier (<a href="http://orcid.org/0000-0002-2085-6629">Félix Chénier</a>)
**Repository:** <a href="https://github.com/felixchenier/kineticstoolkit" target ="_blank">https://github.com/felixchenier/kineticstoolkit</a>
**Version:** 5.0.1
