### 2021-06-08 #3343 @kthyng
@meg-simula is this type of biomechanics closer to your area of expertise?
