### 2021-07-19 #3343 @felixchenier
Good morning @meg-simula
Thanks for being part of this amazing journal.
I'm wondering if my submission has been forgotten since it's been in pre-review for a while, but I understand we are in a pandemic summer so I don't expect things to go fast. I was just wondering.
Thank you again,
Felix Chenier
