### 2021-07-29 #3343 @felixchenier
Good morning @kthyng

I don't know if this is standard since I this is my first submission to JOSS, but my manuscript has been in pre-review for nearly 2 months and I couldn't receive news from the editor (@meg-simula).

If this is standard or to be expected during this pandemic summer, please disregard this message.

Thank you and best regards,

Félix Chénier
