### 2021-07-30 9d154cd
[feature/Changes-in-kinematics-nomenclature-#40] #40 Finalized nomenclature in kinematics. Tutorials and tests to convert to the new nomenclature.
