+++ kineticstoolkit/kinematics.py
@@ -27,7 +27,7 @@ __license__ = "Apache 2.0"
 
 import kineticstoolkit.geometry as geometry
 from kineticstoolkit import TimeSeries
-from kineticstoolkit.decorators import unstable, directory
+from kineticstoolkit.decorators import unstable, directory, deprecated
 from typing import Sequence, Dict, Any
 
 import numpy as np
@@ -351,17 +351,20 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []):
 #            return(the_timeseries)
 
 
-def create_rigid_body_config(
+@deprecated(since='0.5.2', until='1.0',
+            details='It has been replaced by `define_body` because '
+                    'the latter has a clearer name.')
+def create_rigid_body_config(markers, marker_names):
+    """Create a rigid body configuration based on a static acquisition."""
+    return define_body(markers, marker_names)
+
+
+def define_body(
     markers: TimeSeries,
     marker_names: Sequence[str]
 ) -> Dict[str, Any]:
     """
-    Create a rigid body configuration based on a static acquisition.
-
-    Warning
-    -------
-    This function, which has been introduced in 0.4, is still experimental and
-    may change signature or behaviour in the future.
+    Create a generic rigid body definition based on a static acquisition.
 
     Parameters
     ----------
@@ -410,13 +413,29 @@ def create_rigid_body_config(
     }
 
 
-def register_markers(
+@deprecated(since='0.5.2', until='1.0',
+            details='It has been replaced by `track_body` because '
+                    'the latter has a shorter name and performs only one '
+                    'action (which is good).')
+def register_markers(markers, rigid_body_configs, verbose=False):
+    """Calculate the trajectory of rigid bodies."""
+    rigid_bodies = TimeSeries(time=markers.time,
+                              time_info=markers.time_info,
+                              events=markers.events)
+    for rigid_body_name in rigid_body_configs:
+        if verbose is True:
+            print(f'Computing trajectory of rigid body {rigid_body_name}...')
+        rigid_bodies.data[rigid_body_name] = track_body(
+            markers, rigid_body_configs[rigid_body_name])
+    return rigid_bodies
+
+
+def track_body(
         markers: TimeSeries,
-        rigid_body_configs: Dict[str, Dict[str, Any]],
-        verbose: bool = False
-) -> TimeSeries:
+        body_definition: Dict[str, Any],
+) -> np.ndarray:
     """
-    Calculate the trajectory of rigid bodies.
+    Track the trajectory of rigid bodies as a series of frames.
 
     Calculates the trajectory of rigid bodies using
     `ktk.geometry.register_points`.
@@ -430,102 +449,89 @@ def register_markers(
     ----------
     markers
         Markers trajectories to calculate the rigid body trajectories on.
-    rigid_body_configs
-        A dict where each key is a rigid body configuration, and where
-        each rigid body configuration is a dict with the following
-        keys: 'MarkerNames' and 'LocalPoints'.
-    verbose
-        Optional. Set to True to print the rigid body being computed.
+    body_definition
+        A dict with the following keys: 'MarkerNames' and 'LocalPoints' as
+        returned by `ktk.kinematics.define_body()`.
 
     Returns
     -------
-    TimeSeries
-        TimeSeries where each data key is a Nx4x4 series of rigid
-        transformations.
+    Nx4x4 series of frames.
     """
-    rigid_bodies = TimeSeries(time=markers.time,
-                              time_info=markers.time_info,
-                              events=markers.events)
+    # Set local and global points
+    local_points = body_definition['LocalPoints']
 
-    for rigid_body_name in rigid_body_configs:
-        if verbose is True:
-            print(f'Computing trajectory of rigid body {rigid_body_name}...')
+    global_points = np.empty(
[…]
