### 2021-07-30 ab728f6
[#40-second-attempt] #40 I guess it's done, using the new approach. Need to move the new tutorial in the right section, and to remove the FutureWarnings from the unit tests.
