+++ kineticstoolkit/kinematics.py
@@ -351,16 +351,8 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []):
 #            return(the_timeseries)
 
 
-@deprecated(since='0.5.2', until='1.0',
-            details='It has been replaced by `define_body` because '
-                    'the latter has a clearer name.')
-def create_rigid_body_config(markers, marker_names):
-    """Create a rigid body configuration based on a static acquisition."""
-    return define_body(markers, marker_names)
-
-
-def define_body(
-    markers: TimeSeries,
+def define_rigid_body(
+    kinematics: TimeSeries,
     marker_names: Sequence[str]
 ) -> Dict[str, Any]:
     """
@@ -368,22 +360,21 @@ def define_body(
 
     Parameters
     ----------
-    markers
-        Markers trajectories during the static acquisition.
+    kinematics
+        Markers trajectories during a static acquisition.
     marker_names
         The markers that define the rigid body.
 
     Returns
     -------
-    Dict[str, Any]
-        Dictionary with the following keys:
+    Dictionary with the following keys:
 
         - MarkerNames : the same as marker_names
         - LocalPoints : a 1x4xM array that indicates the local position of
           each M marker in the created rigid body config.
 
     """
-    n_samples = len(markers.time)
+    n_samples = len(kinematics.time)
     n_markers = len(marker_names)
 
     # Construct the global points array
@@ -391,7 +382,7 @@ def define_body(
 
     for i_marker, marker in enumerate(marker_names):
         global_points[:, :, i_marker] = \
-            markers.data[marker][:, :]
+            kinematics.data[marker][:, :]
 
     # Remove samples where at least one marker is invisible
     global_points = global_points[~geometry.isnan(global_points)]
@@ -413,102 +404,80 @@ def define_body(
     }
 
 
-@deprecated(since='0.5.2', until='1.0',
-            details='It has been replaced by `track_body` because '
-                    'the latter has a shorter name and performs only one '
-                    'action (which is good).')
-def register_markers(markers, rigid_body_configs, verbose=False):
-    """Calculate the trajectory of rigid bodies."""
-    rigid_bodies = TimeSeries(time=markers.time,
-                              time_info=markers.time_info,
-                              events=markers.events)
-    for rigid_body_name in rigid_body_configs:
-        if verbose is True:
-            print(f'Computing trajectory of rigid body {rigid_body_name}...')
-        rigid_bodies.data[rigid_body_name] = track_body(
-            markers, rigid_body_configs[rigid_body_name])
-    return rigid_bodies
-
-
-def track_body(
-        markers: TimeSeries,
-        body_definition: Dict[str, Any],
-) -> np.ndarray:
+def track_rigid_body(
+        kinematics: TimeSeries,
+        rigid_body_definition: Dict[str, Any],
+        rigid_body_name: str
+) -> TimeSeries:
     """
-    Track the trajectory of rigid bodies as a series of frames.
+    Track a rigid body from markers trajectories.
 
-    Calculates the trajectory of rigid bodies using
-    `ktk.geometry.register_points`.
-
-    Warning
-    -------
-    This function, which has been introduced in 0.4, is still experimental and
-    may change signature or behaviour in the future.
+    This function tracks the specified rigid body in a TimeSeries that
+    contains the required markers, and adds the tracked rigid body to a copy
+    of the input TimeSeries as a Nx4x4 series of frames.
 
     Parameters
     ----------
-    markers
-        Markers trajectories to calculate the rigid body trajectories on.
-    body_definition
+    kinematics
[…]
