### 2021-07-31 eb1d346
[#55-TimeSeries-methods-should-return-copies] #55 Code changed in timeseries.py. Doctests succeed.
