+++ kineticstoolkit/timeseries.py
@@ -522,7 +522,11 @@ class TimeSeries():
         ts.time = dataframe.index.to_numpy()
         return ts
 
-    def add_data_info(self, data_key: str, info_key: str, value: Any) -> None:
+    def add_data_info(
+            self,
+            data_key: str,
+            info_key: str,
+            value: Any) -> 'TimeSeries':
         """
         Add metadata to TimeSeries' data.
 
@@ -537,15 +541,15 @@ class TimeSeries():
 
         Returns
         -------
-        None.
+        A copy of the TimeSeries with the added data info.
 
         Example
         -------
         We create a TimeSeries with some data info:
 
         >>> ts = ktk.TimeSeries()
-        >>> ts.add_data_info('Forces', 'Unit', 'N')
-        >>> ts.add_data_info('Marker1', 'Color', [43, 2, 255])
+        >>> ts = ts.add_data_info('Forces', 'Unit', 'N')
+        >>> ts = ts.add_data_info('Marker1', 'Color', [43, 2, 255])
 
         Let's see the result:
 
@@ -556,12 +560,17 @@ class TimeSeries():
         {'Color': [43, 2, 255]}
 
         """
+        ts = self.copy()
         try:
-            self.data_info[data_key][info_key] = value
+            ts.data_info[data_key][info_key] = value
         except KeyError:
-            self.data_info[data_key] = {info_key: value}
+            ts.data_info[data_key] = {info_key: value}
+        return ts
 
-    def remove_data_info(self, data_key: str, info_key: str) -> None:
+    def remove_data_info(
+            self,
+            data_key: str,
+            info_key: str) -> 'TimeSeries':
         """
         Remove metadata from a TimeSeries' data.
 
@@ -572,6 +581,10 @@ class TimeSeries():
         info_key
             The key of the info dict.
 
+        Returns
+        -------
+        A copy of the TimeSeries with the removed data info.
+
         Note
         ----
         No warning or exception is raised if the data key does not exist.
@@ -579,21 +592,26 @@ class TimeSeries():
         Example
         -------
         >>> ts = ktk.TimeSeries()
-        >>> ts.add_data_info('Forces', 'Unit', 'N')
+        >>> ts = ts.add_data_info('Forces', 'Unit', 'N')
         >>> ts.data_info['Forces']
         {'Unit': 'N'}
 
-        >>> ts.remove_data_info('Forces', 'Unit')
+        >>> ts = ts.remove_data_info('Forces', 'Unit')
         >>> ts.data_info['Forces']
         {}
 
         """
+        ts = self.copy()
         try:
-            self.data_info[data_key].pop(info_key)
+            ts.data_info[data_key].pop(info_key)
         except KeyError:
             pass
+        return ts
 
-    def rename_data(self, old_data_key: str, new_data_key: str) -> None:
+    def rename_data(
+            self,
+            old_data_key: str,
+            new_data_key: str) -> 'TimeSeries':
         """
         Rename a key in data and data_info.
 
@@ -604,11 +622,15 @@ class TimeSeries():
         new_data_key
             New name of the data key.
 
+        Returns
+        -------
+        A copy of the TimeSeries with the renamed data.
+
         Example
         -------
         >>> ts = ktk.TimeSeries()
         >>> ts.data['test'] = np.arange(10)
-        >>> ts.add_data_info('test', 'Unit', 'm')
+        >>> ts = ts.add_data_info('test', 'Unit', 'm')
 
         >>> ts.data
         {'test': array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}
@@ -616,7 +638,7 @@ class TimeSeries():
         >>> ts.data_info
[…]
