### 2021-08-01 0c699d8
[#55-TimeSeries-methods-should-return-copies] #55 All doctests and unit tests pass.
