+++ tests/test_timeseries.py
@@ -69,14 +69,14 @@ def test_TimeSeries():
     ts.time_info['info'] = 'test'
     assert ts == ts2
 
-    ts2.add_data_info('test', 'info', 'test')
+    ts2 = ts2.add_data_info('test', 'info', 'test')
     assert ts != ts2
-    ts.add_data_info('test', 'info', 'test')
+    ts = ts.add_data_info('test', 'info', 'test')
     assert ts == ts2
 
-    ts2.add_event(1, 'one')
+    ts2 = ts2.add_event(1, 'one')
     assert ts != ts2
-    ts.add_event(1, 'one')
+    ts = ts.add_event(1, 'one')
     assert ts == ts2
 
 
@@ -105,19 +105,19 @@ def test_from_dataframe():
 
 def test_add_remove_data_info():
     ts = ktk.TimeSeries()
-    ts.add_data_info('Force', 'Unit', 'N')
-    ts.add_data_info('Force', 'Other', 'hello')
+    ts = ts.add_data_info('Force', 'Unit', 'N')
+    ts = ts.add_data_info('Force', 'Other', 'hello')
 
     # Check that data_info was added
     assert ts.data_info['Force']['Unit'] == 'N'
     assert ts.data_info['Force']['Other'] == 'hello'
 
     # Test removing non-existing data_info
-    ts.remove_data_info('Nonexisting', 'Other')
+    ts = ts.remove_data_info('Nonexisting', 'Other')
     assert len(ts.data_info['Force']) == 2
 
     # Test removing existing data_info
-    ts.remove_data_info('Force', 'Other')
+    ts = ts.remove_data_info('Force', 'Other')
     assert ts.data_info['Force']['Unit'] == 'N'
     assert len(ts.data_info['Force']) == 1
 
@@ -125,29 +125,29 @@ def test_add_remove_data_info():
 def test_rename_remove_data():
     ts = ktk.TimeSeries()
     ts.data['Force'] = np.arange(10)
-    ts.rename_data('Force', 'Moment')
+    ts = ts.rename_data('Force', 'Moment')
     assert np.allclose(ts.data['Moment'], np.arange(10))
 
     # Same test but with data info
-    ts.add_data_info('Moment', 'Unit', 'Nm')
-    ts.rename_data('Moment', 'Power')
+    ts = ts.add_data_info('Moment', 'Unit', 'Nm')
+    ts = ts.rename_data('Moment', 'Power')
     assert np.allclose(ts.data['Power'], np.arange(10))
     assert ts.data_info['Power']['Unit'] == 'Nm'
 
     # Rename inexistent data
-    ts.rename_data('NoKey', 'Anything')
+    ts = ts.rename_data('NoKey', 'Anything')
     assert np.allclose(ts.data['Power'], np.arange(10))
     assert len(ts.data) == 1
     assert len(ts.data_info) == 1
 
     # Remove inexistent data
-    ts.remove_data('NoKey')
+    ts = ts.remove_data('NoKey')
     assert np.allclose(ts.data['Power'], np.arange(10))
     assert len(ts.data) == 1
     assert len(ts.data_info) == 1
 
     # Remove data
-    ts.remove_data('Power')
+    ts = ts.remove_data('Power')
     assert len(ts.data) == 0
     assert len(ts.data_info) == 0
 
@@ -155,9 +155,9 @@ def test_rename_remove_data():
 def test_rename_event():
     # Original doctest
     ts = ktk.TimeSeries()
-    ts.add_event(5.5, 'event1')
[…]
