### 2021-08-01 6935315
[#55-TimeSeries-methods-should-return-copies] Closes #55
