+++ kineticstoolkit/timeseries.py
@@ -541,7 +541,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the added data info.
+        TimeSeries
+            A copy of the TimeSeries with the added data info.
 
         Example
         -------
@@ -583,7 +584,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the removed data info.
+        TimeSeries
+            A copy of the TimeSeries with the removed data info.
 
         Note
         ----
@@ -624,7 +626,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the renamed data.
+        TimeSeries
+            A copy of the TimeSeries with the renamed data.
 
         Example
         -------
@@ -647,7 +650,7 @@ class TimeSeries():
         {'signal': {'Unit': 'm'}}
 
         """
-        ts= self.copy()
+        ts = self.copy()
         try:
             ts.data[new_data_key] = ts.data.pop(old_data_key)
         except KeyError:
@@ -669,7 +672,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the removed data.
+        TimeSeries
+            A copy of the TimeSeries with the removed data.
 
         Note
         ----
@@ -722,7 +726,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the added event.
+        TimeSeries
+            A copy of the TimeSeries with the added event.
 
         Example
         -------
@@ -761,7 +766,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the renamed event.
+        TimeSeries
+            A copy of the TimeSeries with the renamed event.
 
         Example
         -------
@@ -791,7 +797,7 @@ class TimeSeries():
         ts = self.copy()
 
         if old_name == new_name:
-            return ts # Nothing to do.
+            return ts  # Nothing to do.
 
         if occurrence is None:
             # Rename every occurrence of this event
@@ -826,7 +832,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the removed event.
+        TimeSeries
+            A copy of the TimeSeries with the removed event.
 
         Example
         -------
@@ -1066,7 +1073,8 @@ class TimeSeries():
 
         Returns
         -------
-        A copy of the TimeSeries with the sorted events.
+        TimeSeries
+            A copy of the TimeSeries with the sorted events.
 
         Example
         -------
@@ -1268,7 +1276,8 @@ class TimeSeries():
 
         Returns
         -------
-        The index in the time vector.
+        int
+            The index in the time vector.
 
         Example
         -------
@@ -1300,8 +1309,9 @@ class TimeSeries():
 
         Returns
         -------
-        The index in the time vector. If no value is before the specified
-        time, a value of np.nan is returned.
+        int
+            The index in the time vector. If no value is before the specified
+            time, a value of np.nan is returned.
 
         Example
         -------
@@ -1364,8 +1374,9 @@ class TimeSeries():
 
         Returns
         -------
-        The index in the time vector. If no value is after the
-        specified time, a value of np.nan is returned.
+        int
+            The index in the time vector. If no value is after the
+            specified time, a value of np.nan is returned.
 
         Example
         -------
@@ -1430,7 +1441,8 @@ class TimeSeries():
 
         Returns
         -------
-        The index of the event or np.nan if no event found.
+        int
+            The index of the event or np.nan if no event found.
 
         """
         occurrence = int(occurrence)
@@ -1473,8 +1485,9 @@ class TimeSeries():
 
         Returns
         -------
[…]
