### 2021-08-01 b18c6b3
[#40-second-attempt] [STABLE] Solves #58, #56, #40.
