+++ tests/test_kinematics.py
@@ -12,6 +12,7 @@ For now this is the original tutorial without any additional check.
 
 import kineticstoolkit as ktk
 import numpy as np
+import warnings
 
 
 def test_read_n3d_file():
@@ -66,215 +67,111 @@ def test_read_c3d_file():
     assert(markers.data_info['Foot_Marker1']['Unit'] == 'm')
 
 
-def test_reconstruction():
-    """Test the reconstruction chain."""
-    marker_names = ['Probe1', 'Probe2', 'Probe3', 'Probe4', 'Probe5', 'Probe6',
-                    'WheelchairRearR', 'WheelchairRearL', 'WheelchairFront',
-                    'ScapulaL1', 'ScapulaL2', 'ScapulaL3',
-                    'ScapulaR1', 'ScapulaR2', 'ScapulaR3',
-                    'Head1', 'Head2', 'Head3',
-                    'Sternum',
-                    'ArmL1', 'ArmL2', 'ArmL3',
-                    'LateralEpicondyleL', 'ForearmL1', 'ForearmL2', 'ForearmL3',
-                    'NAG',
-                    'GloveL1', 'GloveL2', 'GloveL3',
-                    'ArmR1', 'ArmR2', 'ArmR3',
-                    'LateralEpicondyleR', 'ForearmR1', 'ForearmR2', 'ForearmR3',
-                    'NAR',
-                    'GloveR1', 'GloveR2', 'GloveR3']
+def test_reconstruction_old():
+    """Test the deprecated reconstruction chain - that calls the new one."""
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=FutureWarning)
 
-    # %%
-    """
-    Defining the rigid body configurations using the static acquisitions
-    --------------------------------------------------------------------
-    One of the aims of the static trial is to have a sample where every marker
-    is visible. We use this trial to define the rigid body configuration.
-    A rigid body configuration is a list of markers that form a rigid body,
-    along with their local position in the rigid body's reference frame.
+        marker_names = ['Probe1', 'Probe2', 'Probe3', 'Probe4', 'Probe5', 'Probe6',
+                        'WheelchairRearR', 'WheelchairRearL', 'WheelchairFront',
+                        'ScapulaL1', 'ScapulaL2', 'ScapulaL3',
+                        'ScapulaR1', 'ScapulaR2', 'ScapulaR3',
+                        'Head1', 'Head2', 'Head3',
+                        'Sternum',
+                        'ArmL1', 'ArmL2', 'ArmL3',
+                        'LateralEpicondyleL', 'ForearmL1', 'ForearmL2', 'ForearmL3',
+                        'NAG',
+                        'GloveL1', 'GloveL2', 'GloveL3',
+                        'ArmR1', 'ArmR2', 'ArmR3',
+                        'LateralEpicondyleR', 'ForearmR1', 'ForearmR2', 'ForearmR3',
+                        'NAR',
+                        'GloveR1', 'GloveR2', 'GloveR3']
 
-    For this example, we will create rigid bodies for the markers triads
-    'ArmR' and 'ForearmR'.
-    """
-    config = dict()
-    config['RigidBodies'] = dict()
+        config = dict()
+        config['RigidBodies'] = dict()
 
-    # Read the static trial
-    markers = ktk.kinematics.read_n3d_file(
-        ktk.config.root_folder +
-        '/data/kinematics/sample_static.n3d',
-        labels=marker_names)
+        # Read the static trial
+        markers = ktk.kinematics.read_n3d_file(
+            ktk.config.root_folder +
+            '/data/kinematics/sample_static.n3d',
+            labels=marker_names)
 
-    # Create the rigid body configurations
[…]
