+++ tests/test_timeseries.py
@@ -446,6 +446,47 @@ def test_plot():
     plt.close(fig)
 
 
+def test_copy():
+    """Test the copy method."""
+    ts1 = ktk.TimeSeries()
+    ts1.time = np.linspace(0, 99, 100, endpoint=False)
+    ts1.data['signal1'] = np.random.rand(100, 2)
+    ts1.data['signal2'] = np.random.rand(100, 2)
+    ts1.data['signal3'] = np.random.rand(100, 2)
+    ts1 = ts1.add_data_info('signal1', 'Unit', 'Unit1')
+    ts1 = ts1.add_data_info('signal2', 'Unit', 'Unit2')
+    ts1 = ts1.add_data_info('signal3', 'Unit', 'Unit3')
+    ts1 = ts1.add_event(1.54, 'test_event1')
+    ts1 = ts1.add_event(10.2, 'test_event2')
+    ts1 = ts1.add_event(100, 'test_event3')
+
+    # Standard deep copy
+    ts2 = ts1.copy()
+    assert ts1 == ts2
+
+    # A deep copy without data
+    ts2 = ts1.copy(copy_data=False)
+    assert ts1 != ts2
+    assert np.all(ts2.time == ts1.time)
+    assert ts2.data_info['signal1']['Unit'] == 'Unit1'
+    assert ts2.data_info['signal2']['Unit'] == 'Unit2'
+    assert ts2.data_info['signal3']['Unit'] == 'Unit3'
+    assert ts2.events[0].time == 1.54
+    assert ts2.events[1].time == 10.2
+    assert ts2.events[2].time == 100
+
+    # A deep copy without data_info
+    ts2 = ts1.copy(copy_data_info=False)
+    assert ts1 != ts2
+    assert np.all(ts2.time == ts1.time)
+    assert np.all(ts2.data['signal1'] == ts1.data['signal1'])
+    assert np.all(ts2.data['signal2'] == ts1.data['signal2'])
+    assert np.all(ts2.data['signal3'] == ts1.data['signal3'])
+    assert ts2.events[0].time == 1.54
+    assert ts2.events[1].time == 10.2
+    assert ts2.events[2].time == 100
+
+
 if __name__ == "__main__":
     import pytest
     pytest.main([__file__])
