+++ doc/dev/tutorials.rst
@@ -5,4 +5,3 @@ Tutorials for unreleased features
     :maxdepth: 2
 
     inversedynamics
-    reconstruction
