### 2021-08-02 c506bad
[#40-second-attempt] [STABLE] Closes #56. Closes #40. Closes #58. Ready to be merged into master.
