### 2021-08-09 #3343 @meg-simula
@felixchenier Thanks again for your submission of Kinetics toolkit; it looks well within the JOSS scope.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition,[ this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
