### 2021-08-09 #3343 @felixchenier
Hi @meg-simula

Here is the list you asked for. I wrote it for on a previous submission for Kinetics Toolkit, I'm not **that** fast :-)

alcantarar
ixjlyons
melund

then

CELavecchia
anmuller
BabakNSafa
cpontonn

then

trallard
moorepants
