### 2021-08-09 #3343 @meg-simula
:wave: @moorepants, would you be available to review this submission for JOSS? Feel free to revisit the reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
