### 2021-08-10 cc1f5b2
[master] Added tests for TimeSeries.to_dataframe()
