### 2021-08-10 6a7311e
[anthropometry] Changes to the signature, added a draft of test.
