### 2021-08-11 1253567
[stable] Released 0.6.0
