### 2021-08-13 76fc3d2
[#63-and-#64-changes-to-rigid-body-tracking] Feature completed, I think this looks great for now.
