+++ kineticstoolkit/kinematics.py
@@ -29,7 +29,7 @@ import kineticstoolkit.geometry as geometry
 from kineticstoolkit import TimeSeries
 from kineticstoolkit.decorators import unstable, directory, deprecated
 from os.path import exists
-from typing import Sequence, Dict, Any
+from typing import Sequence, Dict, Any, List, Union
 
 import numpy as np
 import warnings
@@ -38,8 +38,10 @@ import struct  # To unpack data from N3D files
 try:
     import ezc3d
 except ModuleNotFoundError:
-    warnings.warn('Could not load ezc3d. Function kinematics.read_c3d_file '
-                  'will not work.')
+    warnings.warn(
+        'Could not load ezc3d. Function kinematics.read_c3d_file '
+        'will not work.'
+    )
 
 
 def read_c3d_file(filename: str) -> TimeSeries:
@@ -94,7 +96,7 @@ def read_c3d_file(filename: str) -> TimeSeries:
     elif point_unit == 'm':
         point_factor = 1
     else:
-        raise(ValueError("Point unit must be 'm' or 'mm'."))
+        raise (ValueError("Point unit must be 'm' or 'mm'."))
 
     n_labels = len(labels)
     for i_label in range(n_labels):
@@ -104,8 +106,9 @@ def read_c3d_file(filename: str) -> TimeSeries:
         label_name = labels[i_label]
 
         output.data[label_name] = np.array(
-            [point_factor, point_factor, point_factor, 1] *
-            reader['data']['points'][:, i_label, :].T)
+            [point_factor, point_factor, point_factor, 1]
+            * reader['data']['points'][:, i_label, :].T
+        )
 
         output = output.add_data_info(label_name, 'Unit', 'm')
 
@@ -140,8 +143,9 @@ def write_c3d_file(filename: str, markers: TimeSeries) -> None:
     - The point unit in the output c3d is m.
 
     """
-    sample_rate = ((markers.time.shape[0] - 1) /
-                   (markers.time[-1] - markers.time[0]))
+    sample_rate = (markers.time.shape[0] - 1) / (
+        markers.time[-1] - markers.time[0]
+    )
 
     marker_list = []
     marker_data = np.zeros((4, len(markers.data), len(markers.time)))
@@ -216,7 +220,7 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []):
         for i_frame in range(n_frames):
             for i_column in range(n_columns):
                 data = struct.unpack('f', fid.read(4))[0]
-                if (data < -1E25):  # technically, it is -3.697314e+28
+                if data < -1e25:  # technically, it is -3.697314e+28
                     data = np.NaN
                 ndi_array[i_frame, i_column] = data
 
@@ -225,8 +229,10 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []):
 
         # Transformation to a TimeSeries
         ts = TimeSeries(
-            time=np.linspace(0, n_frames / collection_frame_frequency,
-                             n_frames))
+            time=np.linspace(
+                0, n_frames / collection_frame_frequency, n_frames
+            )
+        )
 
         for i_marker in range(n_markers):
             if labels != []:
@@ -234,9 +240,14 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []):
             else:
                 label = f'Marker{i_marker}'
 
-            ts.data[label] = np.block([[
-                ndi_array[:, 3*i_marker:3*i_marker+3],
-                np.ones((n_frames, 1))]])
+            ts.data[label] = np.block(
+                [
+                    [
+                        ndi_array[:, 3 * i_marker : 3 * i_marker + 3],
+                        np.ones((n_frames, 1)),
+                    ]
+                ]
+            )
[…]
