+++ kineticstoolkit/anthropometrics.py
@@ -1,0 +1,212 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# Copyright 2020-2021 Félix Chénier
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Provide tools associated to anthropometric measurements and estimates.
+
+This module is in very early development and everything can still change.
+Please don't use this module in production code.
+
+"""
+
+__author__ = "Félix Chénier"
+__copyright__ = "Copyright (C) 2020-2021 Félix Chénier"
+__email__ = "chenier.felix@uqam.ca"
+__license__ = "Apache 2.0"
+
+
+from kineticstoolkit.decorators import unstable, directory
+from kineticstoolkit import TimeSeries, geometry
+from typing import Dict
+import numpy as np
+
+
+"""
+NOTES WHILE I'M BUILDING IT.
+
+Steps
+1. Get xyz pelvis dimensions by creating a temporary local corodinate system.
+2.
+
+Should I just reconstruct everything with the mothos, since it's documented
+as one article, without deconstructing into many functions?
+
+"""
+
+
+def build_pelvis(
+    markers: TimeSeries,
+    /,
+    *,
+    morphology: str = 'M',
+    flesh_adjustment: bool = True,
+) -> TimeSeries:
+    """
+    Adjust markers for flesh and estimate L5S1, hips and COM and local CS.
+
+    The pelvis local coordinate system is located at LS51, with X anterior,
+    Y up and Z right.
+
+    Parameters
+    ----------
+    markers :
+        TimeSeries that contains the following markers as Nx4 series:
+        - AnteriorSuperiorIliacSpineR
+        - AnteriorSuperiorIliacSpineL
+        - PosteriorSuperiorIliacSpineR
+        - PosteriorSuperiorIliacSpineL
+        - PublicSymphysis
+    morphology : Optional.
+        F' for female, 'M' for midsize male, 'LM' for large male. The default
+        is 'M'.
+    flesh_adjusment : Optional.
+        True to adjust the skin markers to anatomical landmarks, False if the
+        markers are already landmarks. The default is
+        True.
+
+    Returns
+    -------
+    TimeSeries
+        Contains the following data:
+        - AnteriorSuperiorIliacSpineR: Nx4 series.
+        - AnteriorSuperiorIliacSpineL: Nx4 series.
+        - PosteriorSuperiorIliacSpineR: Nx4 series.
+        - PosteriorSuperiorIliacSpineL: Nx4 series.
+        - PublicSymphysis: Nx4 series.
+        - PelvisCOM: Center of mass of the pelvis, as an Nx4 series
+        - PelvisLCS: Local coordinate system of the pelvis, as an Nx4x4 series.
+
+    """
+    # Get the required markers
+    try:
+        rasis = markers.data['AnteriorSuperiorIliacSpineR']
+        lasis = markers.data['AnteriorSuperiorIliacSpineL']
[…]
