### 2021-08-13 4d75749
[#63-and-#64-changes-to-rigid-body-tracking] Simplified yet again the rigid body definitions.
