### 2021-08-14 0473d0c
[anthropometry] [anthropometry] Pelvis rigid body definition seems to work. I think that every flesh margin adjustement in this calculation is useless, I probably need to remove it.
