+++ kineticstoolkit/anthropometry.py
@@ -1,137 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-#
-# Copyright 2020-2021 Félix Chénier
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Provide tools associated to anthropometric measurements and estimates.
-
-This module is in very early development and everything can still change.
-Please don't use this module in production code.
-
-"""
-
-__author__ = "Félix Chénier"
-__copyright__ = "Copyright (C) 2020-2021 Félix Chénier"
-__email__ = "chenier.felix@uqam.ca"
-__license__ = "Apache 2.0"
-
-
-from kineticstoolkit.decorators import unstable, directory
-from kineticstoolkit import TimeSeries, geometry
-from typing import Dict
-import numpy as np
-
-
-"""
-NOTES WHILE I'M BUILDING IT.
-
-Steps
-1. Get xyz pelvis dimensions by creating a temporary local corodinate system.
-2.
-
-Should I just reconstruct everything with the mothos, since it's documented
-as one article, without deconstructing into many functions?
-
-"""
-
-
-@unstable
-def dumas2007(markers: TimeSeries) -> TimeSeries:
-    """
-    Create segment frame series and COM point series using Dumas (2007).
-
-    At the moment, this function follows the method described in
-    Dumas, R., Chèze, L., Verriest, J.-P., 2007. Adjustments to McConville
-    et al. and Young et al. body segment inertial parameters. Journal of
-    Biomechanics 40, 543–553. https://doi.org/10.1016/j.jbiomech.2006.02.013
-
-    Parameters
-    ----------
-    kinematics: A TimeSeries that contains a mix of the following markers:
-        Toe3EndL/R, LateralMalleolusL/R, MedialMalleolusL/R,
-        LateralFemoralEpicondyleL/R, MedialFemoralEpicondyleL/R,
-        AnteriorIliacSpineL/R, PosteriorIliacSpineL/R, PubicSymphysis,
-        GlenohumeralJointL/R,
-        LateralHumeralEpicondyleL/R, MedialHumeralEpicondyleL/R,
-        RadialStyloidL/R, UlnarStyloidL/R,
-        HandMeta2L/R, HandMeta5L/R,
-        C7, JugularNotch, HeadVertex, Sellion.
-
-    Returns
-    -------
-    TimeSeries
-        A TimeSeries that contains these data:
-            Centres of mass (Nx4 series):
-                ComPelvis, ComTorso, ComHeadNeck, ComArmL, ComArmR,
-                ComForearmL, ComForearmR, ComHandL, ComHandR,
-                ComThighL, ComThighR, ComLegL, ComLegR, ComFootL, ComFootR
-            Frames (Nx4x4 series):
-                Pelvis, Torso, HeadNeck, ArmL, ArmR,
-                ForearmL, ForearmR, HandL, HandR,
-                ThighL, ThighR, LegL, LegR, FootL, FootR
-
-    Notes
-    -----
-    If one or many required markers are missing, warnings are raised and the
-    corresponding segments's anthropometrics are ignored.
-
-    """
-    return markers
-    # --- Pelvis
-
[…]
