+++ kineticstoolkit/timeseries.py
@@ -51,7 +51,8 @@ WINDOW_PLACEMENT = {'top': 50, 'right': 0}
 
 
 def dataframe_to_dict_of_arrays(
-        dataframe: pd.DataFrame) -> Dict[str, np.ndarray]:
+    dataframe: pd.DataFrame,
+) -> Dict[str, np.ndarray]:
     """
     Convert a pandas DataFrame to a dict of numpy ndarrays.
 
@@ -118,7 +119,8 @@ def dataframe_to_dict_of_arrays(
         splitted = column.split('[')
         if len(splitted) > 1:  # There are brackets
             new_columns.append(
-                splitted[0] + '[' + splitted[1].replace(' ', ''))
+                splitted[0] + '[' + splitted[1].replace(' ', '')
+            )
         else:
             new_columns.append(column)
     dataframe.columns = columns
@@ -152,17 +154,21 @@ def dataframe_to_dict_of_arrays(
         else:
             highest_dims = np.max(np.array(dimensions[key]), axis=0)
 
-            columns = [key + str(dim).replace(' ', '')
-                       for dim in sorted(dimensions[key])]
+            columns = [
+                key + str(dim).replace(' ', '')
+                for dim in sorted(dimensions[key])
+            ]
             out[key] = dataframe[columns].to_numpy()
-            out[key] = np.reshape(out[key],
-                                  [n_samples] + (highest_dims + 1).tolist())
+            out[key] = np.reshape(
+                out[key], [n_samples] + (highest_dims + 1).tolist()
+            )
 
     return out
 
 
 def dict_of_arrays_to_dataframe(
-        dict_of_arrays: Dict[str, np.ndarray]) -> pd.DataFrame:
+    dict_of_arrays: Dict[str, np.ndarray]
+) -> pd.DataFrame:
     """
     Convert a dict of ndarray of any dimension to a pandas DataFrame.
 
@@ -255,8 +261,9 @@ def dict_of_arrays_to_dataframe(
             # the same way we reshaped the original data. Then we know where
             # the original indices are in the new reshaped data.
             original_indices = np.indices(original_data_shape[1:])
-            reshaped_indices = np.reshape(original_indices,
-                                          (-1, reshaped_data_shape[1]))
+            reshaped_indices = np.reshape(
+                original_indices, (-1, reshaped_data_shape[1])
+            )
 
             # Hint for my future self:
             # For a one-dimension series, reshaped_indices will be:
@@ -281,8 +288,9 @@ def dict_of_arrays_to_dataframe(
 
                     for i_indice in range(0, n_indices):
                         this_column_name += str(
-                            reshaped_indices[i_indice, i_column])
-                        if i_indice == n_indices-1:
+                            reshaped_indices[i_indice, i_column]
+                        )
+                        if i_indice == n_indices - 1:
                             this_column_name += ']'
                         else:
                             this_column_name += ','
@@ -301,7 +309,7 @@ def dict_of_arrays_to_dataframe(
 
 
 @dataclass
-class TimeSeriesEvent():
+class TimeSeriesEvent:
     """
     Define an event in a timeseries.
 
@@ -362,7 +370,7 @@ class TimeSeriesEvent():
         return {'Time': self.time, 'Name': self.name}
 
 
-class TimeSeries():
+class TimeSeries:
     """
     A class that implements TimeSeries.
 
@@ -400,12 +408,13 @@ class TimeSeries():
     """
 
     def __init__(
-            self,
-            time: np.ndarray = np.array([]),
-            time_info: Dict[str, Any] = {'Unit': 's'},
-            data: Dict[str, np.ndarray] = {},
[…]
