### 2021-08-17 faa2185
[master] #55 Added in_place parameters to TimeSeries methods.
