+++ kineticstoolkit/timeseries.py
@@ -51,8 +51,7 @@ WINDOW_PLACEMENT = {'top': 50, 'right': 0}
 
 
 def dataframe_to_dict_of_arrays(
-    dataframe: pd.DataFrame,
-) -> Dict[str, np.ndarray]:
+        dataframe: pd.DataFrame) -> Dict[str, np.ndarray]:
     """
     Convert a pandas DataFrame to a dict of numpy ndarrays.
 
@@ -167,8 +166,7 @@ def dataframe_to_dict_of_arrays(
 
 
 def dict_of_arrays_to_dataframe(
-    dict_of_arrays: Dict[str, np.ndarray]
-) -> pd.DataFrame:
+        dict_of_arrays: Dict[str, np.ndarray]) -> pd.DataFrame:
     """
     Convert a dict of ndarray of any dimension to a pandas DataFrame.
 
@@ -534,8 +532,12 @@ class TimeSeries:
         return ts
 
     def add_data_info(
-        self, data_key: str, info_key: str, value: Any
-    ) -> 'TimeSeries':
+            self,
+            data_key: str,
+            info_key: str,
+            value: Any,
+            *,
+            in_place: bool = False) -> 'TimeSeries':
         """
         Add metadata to TimeSeries' data.
 
@@ -547,11 +549,15 @@ class TimeSeries:
             The key of the info dict.
         value
             The info.
+        in_place
+            Optional. True to modify the original TimeSeries. False
+            to return a modified copy of the TimeSeries while leaving the
+            original TimeSeries intact. The default is False.
 
         Returns
         -------
         TimeSeries
-            A copy of the TimeSeries with the added data info.
+            The TimeSeries with the added data info.
 
         Example
         -------
@@ -570,14 +576,19 @@ class TimeSeries:
         {'Color': [43, 2, 255]}
 
         """
-        ts = self.copy()
+        ts = self if in_place else self.copy()
         try:
             ts.data_info[data_key][info_key] = value
         except KeyError:
             ts.data_info[data_key] = {info_key: value}
         return ts
 
-    def remove_data_info(self, data_key: str, info_key: str) -> 'TimeSeries':
+    def remove_data_info(
+            self,
+            data_key: str,
+            info_key: str,
+            *,
+            in_place: bool = False) -> 'TimeSeries':
         """
         Remove metadata from a TimeSeries' data.
 
@@ -587,11 +598,15 @@ class TimeSeries:
             The data key the info corresponds to.
         info_key
             The key of the info dict.
+        in_place
+            Optional. True to modify the original TimeSeries. False
+            to return a modified copy of the TimeSeries while leaving the
+            original TimeSeries intact. The default is False.
 
         Returns
         -------
         TimeSeries
-            A copy of the TimeSeries with the removed data info.
+            The TimeSeries with the removed data info.
 
         Note
         ----
@@ -609,7 +624,7 @@ class TimeSeries:
         {}
 
         """
-        ts = self.copy()
+        ts = self if in_place else self.copy()
         try:
             ts.data_info[data_key].pop(info_key)
         except KeyError:
@@ -617,8 +632,11 @@ class TimeSeries:
         return ts
 
     def rename_data(
-        self, old_data_key: str, new_data_key: str
-    ) -> 'TimeSeries':
+            self,
+            old_data_key: str,
+            new_data_key: str,
+            *,
+            in_place: bool = False) -> 'TimeSeries':
         """
         Rename a key in data and data_info.
 
@@ -628,11 +646,15 @@ class TimeSeries:
             Name of the current data key.
         new_data_key
             New name of the data key.
+        in_place
+            Optional. True to modify and return the original TimeSeries. False
+            to return a modified copy of the TimeSeries while leaving the
[…]
