### 2021-08-18 5a1065c
[master] #63 and #64. +Removed useless notebook files in documentation.
