### 2021-08-24 2bfad51
[anthropometry] Creation of the pelvis definition.
