+++ kineticstoolkit/anthropometrics.py
@@ -32,9 +32,10 @@ __license__ = "Apache 2.0"
 
 
 from kineticstoolkit.decorators import unstable, directory
-from kineticstoolkit import TimeSeries, geometry
+from kineticstoolkit import TimeSeries, geometry, config
 from typing import Dict
 import numpy as np
+import pandas as pd
 
 
 """
@@ -50,14 +51,10 @@ as one article, without deconstructing into many functions?
 """
 
 
-def define_pelvis_coordinate_system(
-    markers: TimeSeries,
-    /,
-    *,
-    sex: str = 'M',
-) -> Dict[str, np.ndarray]:
+def define_pelvis(
+        markers: TimeSeries, /, *, sex: str = 'M') -> Dict[str, np.ndarray]:
     """
-    Create a segment definition for the Pelvis.
+    Create a local coordinate system for the Pelvis.
 
     The pelvis local coordinate system is located at LS51, with X anterior,
     Y up and Z right. The segment definition is based on
@@ -121,7 +118,8 @@ def define_pelvis_coordinate_system(
     # Create a temporary, well-aligned local coordinate system
     mpsis = 0.5 * (rpsis + lpsis)
     temp_LCS = geometry.create_frames(
-        origin=mpsis, z=(rasis - lasis), xz=(0.5 * (rasis - lasis) - mpsis)
+        origin=mpsis, z=(rasis - lasis),
+        xz=(0.5 * (rasis - lasis) - mpsis),
     )
 
     # Calculate the markers in the local coordinate system
@@ -133,46 +131,22 @@ def define_pelvis_coordinate_system(
 
     # Create a rigid body definition using these locations
     pelvis_definition = {
-        'AnteriorSuperiorIliacSpineR': np.nanmean(local_rasis, axis=0)[
-            np.newaxis
-        ],
-        'AnteriorSuperiorIliacSpineL': np.nanmean(local_lasis, axis=0)[
-            np.newaxis
-        ],
-        'PosteriorSuperiorIliacSpineR': np.nanmean(local_rpsis, axis=0)[
-            np.newaxis
-        ],
-        'PosteriorSuperiorIliacSpineL': np.nanmean(local_lpsis, axis=0)[
-            np.newaxis
-        ],
-        'PubicSymphysis': np.nanmean(local_sym, axis=0)[np.newaxis],
+        'AnteriorSuperiorIliacSpineR':
+            np.nanmean(local_rasis, axis=0)[np.newaxis],
+        'AnteriorSuperiorIliacSpineL':
+            np.nanmean(local_lasis, axis=0)[np.newaxis],
+        'PosteriorSuperiorIliacSpineR':
+            np.nanmean(local_rpsis, axis=0)[np.newaxis],
+        'PosteriorSuperiorIliacSpineL':
+            np.nanmean(local_lpsis, axis=0)[np.newaxis],
+        'PubicSymphysis':
+            np.nanmean(local_sym, axis=0)[np.newaxis],
     }
 
-    # Flesh margin adjustment to calculate pelvis width
-    if flesh_adjustment:
-        adjusted_local_rasis = local_rasis + np.array([[-10, 0, 0, 0]]) / 1000
-        adjusted_local_lasis = local_lasis + np.array([[-10, 0, 0, 0]]) / 1000
-        adjusted_local_sym = (
-            local_sym + np.array([[-17.7, -17.7, 0, 0]]) / 1000
-        )
-    else:
-        adjusted_local_rasis = local_rasis
-        adjusted_local_lasis = local_lasis
-        adjusted_local_sym = local_sym
[…]
