### 2021-08-25 089b02b
[anthropometry] Added marker creation for the trunk, arms and forearms.
