+++ kineticstoolkit/anthropometrics.py
@@ -32,10 +32,11 @@ __license__ = "Apache 2.0"
 
 
 from kineticstoolkit.decorators import unstable, directory
-from kineticstoolkit import TimeSeries, geometry, config
+from kineticstoolkit import TimeSeries, geometry, config, kinematics, Player
 from typing import Dict
 import numpy as np
 import pandas as pd
+from warnings import warn
 
 
 """
@@ -51,14 +52,264 @@ as one article, without deconstructing into many functions?
 """
 
 
-def define_pelvis(
-        markers: TimeSeries, /, *, sex: str = 'M') -> Dict[str, np.ndarray]:
+def _add_l5s1_hip_joints(
+        markers: TimeSeries,
+        sex: str = 'M') -> None:
+    """Add L5S1 and hip joint centers based on Reed et al. (1999)."""
+    # Get the required markers
+    try:
+        rasis = markers.data['AnteriorSuperiorIliacSpineR']
+        lasis = markers.data['AnteriorSuperiorIliacSpineL']
+        rpsis = markers.data['PosteriorSuperiorIliacSpineR']
+        lpsis = markers.data['PosteriorSuperiorIliacSpineL']
+        sym = markers.data['PubicSymphysis']
+    except KeyError:
+        warn("Not enough markers to reconstruct L5S1 and hip joint centers.")
+        return
+
+    # Create a local coordinate system at the anterior superior iliac spines
+    # midpoint, according to Reed et al.
+    masis = 0.5 * (rasis + lasis)
+    lcs = geometry.create_frames(
+        origin=masis,
+        y=lasis - rasis,
+        yz=masis - sym
+    )
+
+    # Calculate the markers in the local coordinate system
+    local_rasis = geometry.get_local_coordinates(rasis, lcs)
+    local_lasis = geometry.get_local_coordinates(lasis, lcs)
+    local_rpsis = geometry.get_local_coordinates(rpsis, lcs)
+    local_lpsis = geometry.get_local_coordinates(lpsis, lcs)
+    local_sym = geometry.get_local_coordinates(sym, lcs)
+
+    # Create a rigid body definition using these locations
+    pelvis_definition = {
+        'AnteriorSuperiorIliacSpineR':
+            np.nanmean(local_rasis, axis=0)[np.newaxis],
+        'AnteriorSuperiorIliacSpineL':
+            np.nanmean(local_lasis, axis=0)[np.newaxis],
+        'PosteriorSuperiorIliacSpineR':
+            np.nanmean(local_rpsis, axis=0)[np.newaxis],
+        'PosteriorSuperiorIliacSpineL':
+            np.nanmean(local_lpsis, axis=0)[np.newaxis],
+        'PubicSymphysis':
+            np.nanmean(local_sym, axis=0)[np.newaxis],
+    }
+
+    # Track the pelvis using this definition
+    tracked_pelvis = kinematics.track_rigid_body(
+        markers, pelvis_definition, 'Pelvis'
+    )
+
+    # Pelvis width
+    pw = np.abs(np.nanmean(
+        local_rasis[:, 1] - local_lasis[:, 1]
+    ))
+
+    # Create L5S1 if not already included in markers
+    if 'L5S1' not in markers.data:
+        if sex == 'F':
+            local_position = np.array(
+                [[-0.289 * pw, 0.0, 0.172 * pw, 1.0]]
+            )
+        else:  # M
+            local_position = np.array(
+                [[-0.264 * pw, 0.0, 0.126 * pw, 1.0]]
+            )
+        markers.data['L5S1'] = geometry.get_global_coordinates(
+            local_position, tracked_pelvis.data['Pelvis']
[…]
