### 2021-08-26 bdfac06
[anthropometry] Marker infer methods now public.
