+++ kineticstoolkit/anthropometrics.py
@@ -23,6 +23,37 @@ Warning
 This module is in very early development and everything can still change.
 Please don't use this module in production code.
 
+Nomenclature
+------------
+- AnteriorSuperiorIliacSpineL/R
+- PosteriorSuperiorIliacSpineL/R
+- PubicSymphysis
+- L5S1JointCenter
+- C7
+- C7T1JointCenter
+- Suprasternale
+- AcromionL/R
+- GlenohumeralJointCenterL/R
+- LateralHumeralEpicondyleL/R
+- MedialHumeralEpicondyleL/R
+- ElbowJointCenterL/R
+- UlnarStyloidL/R
+- RadialStyloidL/R
+- WristJointCenterL/R
+- CarpalMetaHead1-5
+- FingerTip1-5L/R
+- HipJointCenterL/R
+- LateralFemoralEpicondyleL/R
+- MedialFemoralEpicondyleL/R
+- KneeJointCenterL/R
+- LateralMalleolusL/R
+- MedialMalleolusL/R
+- AnkleJointCenterL/R
+- CalcaneusL/R
+- NavicularL/R
+- TarsalMetaHead1-5L/R
+- ToeTip1-5L/R
+
 """
 
 __author__ = "Félix Chénier"
@@ -39,23 +70,38 @@ import pandas as pd
 from warnings import warn
 
 
-"""
-NOTES WHILE I'M BUILDING IT.
-
-Steps
-1. Get xyz pelvis dimensions by creating a temporary local corodinate system.
-2.
-
-Should I just reconstruct everything with the mothos, since it's documented
-as one article, without deconstructing into many functions?
-
-"""
-
-
-def _add_l5s1_hip_joints(
+@unstable
+def infer_hip_l5s1_joint_centers(
         markers: TimeSeries,
-        sex: str = 'M') -> None:
-    """Add L5S1 and hip joint centers based on Reed et al. (1999)."""
+        sex: str = 'M') -> TimeSeries:
+    """
+    Infer L5S1 and hip joint centers based on anthropometric data.
+
+    Creates L5S1JointCenter, HipJointCenterL and HipJointCenterR based on
+    AnteriorSuperiorIliacSpineL/R, PosteriorSuperiorIliacSpineL/R and
+    PubicSymphysis, following the method presented in Reed et al. (1999)[1]_.
+
+    Parameters
+    ----------
+    markers
+        TimeSeries that contain the trajectory of
+        AnteriorSuperiorIliacSpineL/R, PosteriorSuperiorIliacSpineL/R and
+        PubicSymphysis as Nx4 series.
+    sex
+        Optional. Either 'M' or 'F'. The default is 'M'.
+
+    Returns
+    -------
+    TimeSeries
+        A TimeSeries with L5S1JointCenter, HipJointCenterL and
+        HipJointCenterR as Nx4 series.
+
+    .. [1] Reed, M., Manary, M.A., Schneider, L.W., 1999.
+       Methods for Measuring and Representing Automobile Occupant Posture.
+       Presented at the International Congress &  Exposition, pp. 1999-01–0959.
+       https://doi.org/10.4271/1999-01-0959
+
+    """
     # Get the required markers
     try:
         rasis = markers.data['AnteriorSuperiorIliacSpineR']
@@ -107,54 +153,88 @@ def _add_l5s1_hip_joints(
         local_rasis[:, 1] - local_lasis[:, 1]
     ))
 
-    # Create L5S1 if not already included in markers
-    if 'L5S1' not in markers.data:
-        if sex == 'F':
-            local_position = np.array(
-                [[-0.289 * pw, 0.0, 0.172 * pw, 1.0]]
-            )
-        else:  # M
-            local_position = np.array(
[…]
