### 2021-08-26 b4c11c3
[anthropometry] Can now define every segment's local coordinate system using standardized markers.
