+++ kineticstoolkit/anthropometrics.py
@@ -25,34 +25,42 @@ Please don't use this module in production code.
 
 Nomenclature
 ------------
-- AnteriorSuperiorIliacSpineL/R
-- PosteriorSuperiorIliacSpineL/R
-- PubicSymphysis
-- L5S1JointCenter
+
+Skin markers (either real or reconstructed):
+
+- HeadVertex
+- Sellion
 - C7
-- C7T1JointCenter
 - Suprasternale
-- AcromionL/R
-- GlenohumeralJointCenterL/R
-- LateralHumeralEpicondyleL/R
-- MedialHumeralEpicondyleL/R
-- ElbowJointCenterL/R
-- UlnarStyloidL/R
-- RadialStyloidL/R
-- WristJointCenterL/R
-- CarpalMetaHead1-5
-- FingerTip1-5L/R
-- HipJointCenterL/R
-- LateralFemoralEpicondyleL/R
-- MedialFemoralEpicondyleL/R
-- KneeJointCenterL/R
-- LateralMalleolusL/R
-- MedialMalleolusL/R
-- AnkleJointCenterL/R
-- CalcaneusL/R
-- NavicularL/R
-- TarsalMetaHead1-5L/R
-- ToeTip1-5L/R
+- PubicSymphysis
+- AnteriorSuperiorIliacSpineR/L
+- PosteriorSuperiorIliacSpineR/L
+- AcromionR/L
+- LateralHumeralEpicondyleR/L
+- MedialHumeralEpicondyleR/L
+- UlnarStyloidR/L
+- RadialStyloidR/L
+- CarpalMetaHead2R/L
+- CarpalMetaHead5R/L
+- HipJointCenterR/L
+- LateralFemoralEpicondyleR/L
+- MedialFemoralEpicondyleR/L
+- LateralMalleolusR/L
+- MedialMalleolusR/L
+- CalcaneusR/L
+- NavicularR/L
+- TarsalMetaHead1R/L
+- TarsalMetaHead5R/L
+
+Inferred joint centers:
+
+- L5S1JointCenter
+- C7T1JointCenter
+- GlenohumeralJointCenterR/L
+- ElbowJointCenterR/L
+- WristJointCenterR/L
+- KneeJointCenterR/L
+- AnkleJointCenterR/L
 
 """
 
@@ -71,21 +79,21 @@ from warnings import warn
 
 
 @unstable
-def infer_hip_l5s1_joint_centers(
+def infer_pelvis_joint_centers(
         markers: TimeSeries,
         sex: str = 'M') -> TimeSeries:
     """
     Infer L5S1 and hip joint centers based on anthropometric data.
 
     Creates L5S1JointCenter, HipJointCenterL and HipJointCenterR based on
-    AnteriorSuperiorIliacSpineL/R, PosteriorSuperiorIliacSpineL/R and
+    AnteriorSuperiorIliacSpineR/L, PosteriorSuperiorIliacSpineR/L and
     PubicSymphysis, following the method presented in Reed et al. (1999)[1]_.
 
     Parameters
     ----------
     markers
         TimeSeries that contain the trajectory of
-        AnteriorSuperiorIliacSpineL/R, PosteriorSuperiorIliacSpineL/R and
+        AnteriorSuperiorIliacSpineR/L, PosteriorSuperiorIliacSpineR/L and
         PubicSymphysis as Nx4 series.
     sex
         Optional. Either 'M' or 'F'. The default is 'M'.
@@ -167,7 +175,7 @@ def infer_hip_l5s1_joint_centers(
     else:
         raise ValueError("sex must be either 'M' or 'F'")
 
-    output.data['L5S1'] = geometry.get_global_coordinates(
+    output.data['L5S1JointCenter'] = geometry.get_global_coordinates(
[…]
