### 2021-08-27 1ce964a
[anthropometry] Improvements in anthropometrics and kinematics (#63)
