### 2021-08-31 9b37c9b
[anthropometry] Can now reconstruct every marker and standardized local coordinate systems. Also completed #63.
