+++ kineticstoolkit/anthropometrics.py
@@ -72,44 +72,101 @@ __license__ = "Apache 2.0"
 
 from kineticstoolkit.decorators import unstable, directory
 from kineticstoolkit import TimeSeries, geometry, config, kinematics, Player
-from typing import Dict
+from typing import Dict, Union, Sequence
 import numpy as np
 import pandas as pd
 from warnings import warn
 
 
 @unstable
-def infer_pelvis_joint_centers(
+def infer_joint_centers(
         markers: TimeSeries,
-        sex: str = 'M') -> TimeSeries:
+        /,
+        segment: str,
+        *,
+        sex: str = 'M',
+) -> TimeSeries:
     """
-    Infer L5S1 and hip joint centers based on anthropometric data.
+    Infer joint centers based on anthropometric data.
 
+    This function is aimed to be used on static acquisitions in quasi-neutral
+    position.
+
+    For the pelvis:
     Creates L5S1JointCenter, HipJointCenterL and HipJointCenterR based on
     AnteriorSuperiorIliacSpineR/L, PosteriorSuperiorIliacSpineR/L and
     PubicSymphysis, following the method presented in Reed et al. (1999)[1]_.
 
+    For the trunk:
+    Creates C7T1JointCenter, GlenohumeralJointCenterL and
+    GlenohumeralJointCenterR based on C7, L5S1JointCenter, Suprasternal,
+    AcromionR and AcromionL, following the method presented in Reed et al.
+    (1999)[1]_. The scapulo-thoracic joint must be in a neutral
+    position.
+
+    For the extremities, creates:
+    - ElbowJointCenterR/L as the midpoint between
+      LateralHumeralEpicondyleR/L and MedialHumeralEpicondyleR;
+    - WristJointCenterR as the midpoint between
+      UlnarStyloidR/L and RadialStyloidR/L;
+    - KneeJointCenterR/L as the midpoint between
+      LateralFemoralEpicondyleR/L and MedialFemoralEpicondyleR/L;
+    - AnkleJointCenterR/L as the midpoint between
+      LateralMalleolusR/L and MedialMalleolusR/L.
+
     Parameters
     ----------
     markers
-        TimeSeries that contain the trajectory of
-        AnteriorSuperiorIliacSpineR/L, PosteriorSuperiorIliacSpineR/L and
-        PubicSymphysis as Nx4 series.
+        TimeSeries that contain the trajectory of the required markers as
+        Nx4 series.
+    segment
+        Can be either 'Pelvis', 'Trunk', or 'Extremities'. An error is raised
+        in case of missing markers.
     sex
-        Optional. Either 'M' or 'F'. The default is 'M'.
+        Optional. Can be either 'M' or 'F'. The default is 'M'.
 
     Returns
     -------
     TimeSeries
-        A TimeSeries with L5S1JointCenter, HipJointCenterL and
-        HipJointCenterR as Nx4 series.
+        A TimeSeries that contains the inferred joint center trajectories.
 
+    References
+    ----------
     .. [1] Reed, M., Manary, M.A., Schneider, L.W., 1999.
        Methods for Measuring and Representing Automobile Occupant Posture.
        Presented at the International Congress &  Exposition, pp. 1999-01–0959.
        https://doi.org/10.4271/1999-01-0959
 
     """
+    if segment == 'Pelvis':
+        return _infer_pelvis_joint_centers(markers, sex=sex)
+    elif segment == 'Thorax':
+        return _infer_thorax_joint_centers(markers, sex=sex)
+    elif segment == 'Extremities':
+        return _infer_extremity_joint_centers(markers)
+    elif segment == '':
+        out = markers.copy()
+        for segment in ['Pelvis', 'Thorax', 'Extremities']:
+            try:
+                out.merge(
[…]
