+++ kineticstoolkit/kinematics.py
@@ -362,33 +362,6 @@ def read_n3d_file(filename: str, labels: Sequence[str] = []) -> TimeSeries:
 
 
 @unstable
-@deprecated(since='master', until='master', details="Use create_cluster().")
-def define_rigid_body(
-        kinematics: TimeSeries,
-        marker_names: Sequence[str]) -> Dict[str, np.ndarray]:
-    """
-    Create a generic rigid body definition based on a static acquisition.
-
-    Parameters
-    ----------
-    kinematics
-        Markers trajectories during a static acquisition.
-    marker_names
-        The markers that define the rigid body.
-
-    Returns
-    -------
-    Dict
-        Dictionary where each entry represents the local position of a point
-        (e.g., marker name). The key is the name of the point, the value is a
-        1x4 array that indicates the local position of this point in the rigid
-        body's local coordinate system.
-
-    """
-    return create_cluster(kinematics, marker_names)
-
-
-@unstable
 def create_cluster(
         markers: TimeSeries,
         /,
@@ -442,98 +415,6 @@ def create_cluster(
 
 
 @unstable
-def fit_cluster(
-        markers: TimeSeries,
-        /,
-        cluster: Dict[str, np.ndarray]) -> TimeSeries:
-    """
-    Apply a cluster to a TimeSeries of markers.
-
-    Fits a cluster to a TimeSeries and reconstructs a solidified version of
-    all the points defined in this cluster.
-
-    Parameters
-    ----------
-    markers
-        A TimeSeries that contains point trajectories as Nx4 arrays.
-    cluster
-        A cluster definition as returned by ktk.kinematics.create_cluster().
-
-    Returns
-    -------
-    TimeSeries
-        A TimeSeries with the trajectories of all points defined in the
-        cluster after fitting the input TimeSeries.
-
-    """
-    out = markers.copy(copy_data=False, copy_data_info=False)
-
-    # Track the cluster
-    frames = _track_cluster(markers, cluster)
-
-    for marker in cluster:
-        out.data[marker] = geometry.get_global_coordinates(
-            cluster[marker], frames)
-
-    return out
-
-
-def _track_clusters(
-        markers: TimeSeries,
-        /,
-        clusters: Dict[str, Dict[str, np.ndarray]]) -> TimeSeries:
-    """
-    Track a point cluster in a point TimeSeries.
-
-    Parameters
-    ----------
-    markers
-        TimeSeries with point trajectories from the cluster definition.
-    clusters
-        A dictionary of clusters to track, where each primary key corresponds
-        to one cluster.
-
-    Returns
-    -------
-    TimeSeries
-        A TimeSeries with the tracked clusters.
-
-    """
-    ts = markers.copy(copy_data=False, copy_data_info=False)
-    for cluster in clusters:
-        ts.data[cluster] = _track_cluster(markers, clusters[cluster])
-    return ts
-
-
-def _track_cluster(
-        markers: TimeSeries,
-        cluster: Dict[str, np.ndarray]) -> np.ndarray:
-    """Track one cluster."""
-    # Set local and global points
-    marker_names = cluster.keys()
-    stacked_local_points = np.dstack(
-        [cluster[_] for _ in marker_names])
-
-    global_points = np.empty(
-        (len(markers.time), 4, stacked_local_points.shape[2])
-    )
-
-    for i_marker, marker_name in enumerate(marker_names):
-        if marker_name in markers.data:
-            global_points[:, :, i_marker] = markers.data[marker_name]
-        else:
-            global_points[:, :, i_marker] = np.nan
-
-    (stacked_local_points, global_points) = geometry._match_size(
-        stacked_local_points, global_points
-    )
-
-    # Track the cluster
[…]

[…]
