### 2021-08-31 03a1ed3
[anthropometry] Module anthropometrics now created, with a draft of tutorial and a draft of unit test.
