### 2021-09-08 #3343 @meg-simula
:wave: @ixjlyons or @melund, would either of you be available to review this submission for JOSS? Feel free to revisit the reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html .
