### 2021-09-10 #3714 @meg-simula
The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time. Also note that the current paper is in the branch JOSS, so use e.g.

-> @whedon generate pdf from branch JOSS
-> @whedon check references from branch JOSS

Please feel free to ping me (@meg-simula ) if you have any questions/concerns.
