### 2021-09-11 8ca3498
#31 Added some tests, and removed some untestable functions from unit test coverage.
