### 2021-09-13 felixchenier/kineticstoolkit#66 @alcantarar
JOSS Paper Comments
- Line 23: I suggest using "OpenSim" instead of "Opensim" to match the format of your citation.
- Line 34: Not particularly necessary, but the statement that "python is one of the easiest programming languages to learn" may benefit from a supporting citation.
