### 2021-09-13 9a3e16e
[stable] Enhanced documentation for TimeSeries in website.
