+++ kineticstoolkit/dev.py
@@ -212,11 +212,8 @@ def build_website(clean: bool = False) -> None:  # pragma: no cover
             + '/../kineticstoolkit_doc/stable'
         )
 
-    try:
-        os.mkdir(doc_folder)
-    except Exception:
-        pass
-
+    shutil.rmtree(doc_folder)
+    os.makedirs(doc_folder, exist_ok=True)
     os.rename(
         kineticstoolkit.config.root_folder + '/doc/_build/html',
         doc_folder,
