### 2021-09-14 c87772e
Corrected parameter initialisation error in TimeSeries.ui_edit_events()
