### 2021-09-27 #3714 @alcantarar
👋🏾 @meg-simula I've completed my review. The repo looks good to me and any bugs I've found have been addressed except for [1 typo in the paper](https://github.com/felixchenier/kineticstoolkit/issues/66). I think this toolkit is a nice way to introduce beginner programmers to biomechanical analyses and research with tutorials on how to use the included functions to process a variety of timeseries data. It addresses the valid statement of need outlined in the paper.

@felixchenier, great work! 👍🏾 I particularly enjoyed the 2d kinematics tutorial and was impressed with the 3-D marker/segment visualization. Thanks for developing and maintaining this tool for the biomechanics community!
