### 2021-10-01 #3714 @melund
@meg-simula  Sorry about the delay. I might have been too optimistic about my time.

 I still need to go through the documentation and really try out the functionality. I only just started on that part. I have no time this weekend, but I will try to do it some evening next week.
