+++ joss/JOSS.bib
@@ -1,206 +1,205 @@
 
-@article{wu_isb_2005,
-	title = {{ISB} recommendation on definitions of joint coordinate systems of various joints for the reporting of human joint motion - {Part} {II}: shoulder, elbow, wrist and hand},
+@article{wu_isbrecommendationdefinitions_2005,
+	title = {{ISB} recommendation on definitions of joint coordinate systems of various joints for the reporting of human joint motion - Part {II}: shoulder, elbow, wrist and hand},
 	volume = {38},
 	issn = {00219290},
 	doi = {10.1016/j.jbiomech.2004.05.042},
-	abstract = {In this communication, the Standardization and Terminology Committee (STC) of the International Society of Biomechanics proposes a definition of a joint coordinate system (JCS) for the shoulder, elbow, wrist, and hand. For each joint, a standard for the local axis system in each articulating segment or bone is generated. These axes then standardize the JCS. The STC is publishing these recommendations so as to encourage their use, to stimulate feedback and discussion, and to facilitate further revisions. Adopting these standards will lead to better communication among researchers and clinicians. © 2004 Elsevier Ltd. All rights reserved.},
-	number = {5},
-	journal = {Journal of Biomechanics},
-	author = {Wu, Ge and Van Der Helm, Frans C T and Veeger, H. E J DirkJan and Makhsous, Mohsen and Van Roy, Peter and Anglin, Carolyn and Nagels, Jochem and Karduna, Andrew R and McQuade, Kevin and Wang, Xuguang and Werner, Frederick W and Buchholz, Bryan and {Others}},
-	year = {2005},
-	pmid = {15844264},
-	keywords = {Elbow, Hand, Joint coordinate system, Shoulder, Wrist},
+	abstract = {In this communication, the Standardization and Terminology Committee ({STC}) of the International Society of Biomechanics proposes a definition of a joint coordinate system ({JCS}) for the shoulder, elbow, wrist, and hand. For each joint, a standard for the local axis system in each articulating segment or bone is generated. These axes then standardize the {JCS}. The {STC} is publishing these recommendations so as to encourage their use, to stimulate feedback and discussion, and to facilitate further revisions. Adopting these standards will lead to better communication among researchers and clinicians. © 2004 Elsevier Ltd. All rights reserved.},
 	pages = {981--992},
+	number = {5},
+	journaltitle = {Journal of Biomechanics},
+	author = {Wu, Ge and Van Der Helm, Frans C T and Veeger, H. E J DirkJan and Makhsous, Mohsen and Van Roy, Peter and Anglin, Carolyn and Nagels, Jochem and Karduna, Andrew R and McQuade, Kevin and Wang, Xuguang and Werner, Frederick W and Buchholz, Bryan and {Others}},
+	date = {2005},
+	pmid = {15844264},
+	keywords = {Shoulder, Elbow, Hand, Joint coordinate system, Wrist},
 	file = {Wu et al_2005_ISB recommendation on definitions of joint coordinate systems of various joints.pdf:/Users/felix/Zotero/storage/HNVERNGE/Wu et al_2005_ISB recommendation on definitions of joint coordinate systems of various joints.pdf:application/pdf}
 }
 
-@article{martinez_pyomeca_2020,
-	title = {pyomeca: {An} {Open}-{Source} {Framework} for {Biomechanical} {Analysis}},
+@article{martinez_pyomecaopensourceframework_2020,
+	title = {pyomeca: An Open-Source Framework for Biomechanical Analysis},
 	volume = {5},
 	issn = {2475-9066},
-	shorttitle = {pyomeca},
 	url = {https://joss.theoj.org/papers/10.21105/joss.02431},
[…]
