### 2021-10-12 c26855a
[JOSS] Typos.
