### 2021-10-21 #3714 @felixchenier
@meg-simula

[X] Make a tagged release of your software, and list the version tag of the archived version here.

0.6.2

[X] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
[X] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
[X] Please list the DOI of the archived version here.

10.5281/zenodo.5590654
