### 2021-10-23 1ec6255
Added link to JOSS in README
