+++ README.md
@@ -4,4 +4,4 @@ An Open-Source Python Package to Facilitate Research in Biomechanics
 
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.03714/status.svg)](https://doi.org/10.21105/joss.03714)
 
-Please consult [https://kineticstoolkit.uqam.ca]() for information.
+Please consult https://kineticstoolkit.uqam.ca for information.
