### 2021-11-09 2816cb2
Closes #41 - kinematics.write_c3d_file
