### 2021-11-19 9143e6b
[master] Added a function to calculate the global center of mass in ktk.anthropometrics. Also I switched from nbsphinx to myst_nb to include mermaid figures in future tutorials.
