+++ kineticstoolkit/anthropometrics.py
@@ -97,7 +97,7 @@ def infer_joint_centers(
     AnteriorSuperiorIliacSpineR/L, PosteriorSuperiorIliacSpineR/L and
     PubicSymphysis, following the method presented in Reed et al. (1999)[1]_.
 
-    For the trunk:
+    For the thorax:
     Creates C7T1JointCenter, GlenohumeralJointCenterL and
     GlenohumeralJointCenterR based on C7, L5S1JointCenter, Suprasternal,
     AcromionR and AcromionL, following the method presented in Reed et al.
@@ -120,7 +120,7 @@ def infer_joint_centers(
         TimeSeries that contain the trajectory of the required markers as
         Nx4 series.
     segment
-        Can be either 'Pelvis', 'Trunk', or 'Extremities'. An error is raised
+        Can be either 'Pelvis', 'Thorax', or 'Extremities'. An error is raised
         in case of missing markers.
     sex
         Optional. Can be either 'M' or 'F'. The default is 'M'.
@@ -150,7 +150,7 @@ def infer_joint_centers(
             try:
                 out.merge(
                     infer_joint_centers(
-                        markers,
+                        out,
                         segment=segment,
                         sex=sex,
                     ),
@@ -432,7 +432,8 @@ def track_local_coordinate_systems(
     segments
         Name of the segment. Can be 'Pelvis', 'Thorax', 'HeadNeck',
         'ArmR', 'ArmL', 'ForearmR', 'ForearmL', 'HandR', 'HandL', 'ThighR',
-        'ThighL', 'LegR', 'LegL', 'FootR', or 'FootL'.
+        'ThighL', 'LegR', 'LegL', 'FootR', or 'FootL'. An empty string or
+        an empty list will try to process every segment.
 
     method
         Optional. Select one ot both methods to reconstruct the axial
@@ -449,6 +450,14 @@ def track_local_coordinate_systems(
 
     """
     output = markers.copy(copy_data=False, copy_data_info=False)
+
+    # If we have no segment or an empty list, go with every segment
+    if len(segments) == 0:
+        segments = [
+            'Pelvis', 'Thorax', 'HeadNeck', 'ArmR', 'ArmL', 'ForearmR',
+            'ForearmL', 'HandR', 'HandL', 'ThighR', 'ThighL', 'LegR', 'LegL',
+            'FootR', 'FootL',
+        ]
 
     # If we have a list of segments
     if not isinstance(segments, str):
@@ -688,7 +697,8 @@ def estimate_center_of_mass(
     segments
         Name of the segment. Can be 'Pelvis', 'Thorax', 'HeadNeck',
         'ArmR', 'ArmL', 'ForearmR', 'ForearmL', 'HandR', 'HandL', 'ThighR',
-        'ThighL', 'LegR', 'LegL', 'FootR', or 'FootL'.
+        'ThighL', 'LegR', 'LegL', 'FootR', or 'FootL'. An empty string or
+        an empty list will try to process every segment.
 
     sex
         Optional. Either 'M' or 'F'. The default is 'M'.
@@ -702,6 +712,14 @@ def estimate_center_of_mass(
     """
     markers = markers.copy()  # We will add markers to it so copy it first.
     output = markers.copy(copy_data=False, copy_data_info=False)
+
+    # If we have no segment or an empty list, go with every segment
+    if len(segments) == 0:
+        segments = [
[…]
