### 2021-11-20 cf3930a
[doc_to_md] Transfer from notebooks to myst md files. Dev code is not adapted yet.
