+++ doc/conventions.md
@@ -1,0 +1,29 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+# Conventions
+
+## SI Units
+
+Unless specified, every quantities are expressed in International System units (SI):
+
+- Time: second (s)
+- Length: meter (m)
+- Angle: radian (rad)
+- Mass: kilogram (kg)
+- Force: newton (N)
+- Moment: newton-meter (Nm)
+- Power: watt (W)
+- Energy: joule (J)
+
+## Time as first dimension
+
+Unless specified, when a numpy array is used to express a quantity that varies over time, the time corresponds to the first dimension of the array.
