+++ doc/dev/tutorials.rst
@@ -4,6 +4,6 @@ Tutorials for unreleased features
 .. toctree::
     :maxdepth: 2
 
-    reconstruction
-    anthropometrics
-    inversedynamics
+    01_reconstruction
+    02_anthropometrics
+    03_inversedynamics
