+++ doc/dev/03_inversedynamics.md
@@ -1,0 +1,353 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+# Inverse Dynamics
+
+The [inversedynamics](../api/kineticstoolkit.inversedynamics.rst) module calculates the proximal forces and moments based on the segment's kinematics and distal forces and moments, based on the general method published in Dumas et al., 2004 [^1].
+
+[^1]: R. Dumas, R. Aissaoui, and J. A. De Guise, "A 3D generic inverse dynamic method using wrench notation and quaternion algebra," Comput Meth Biomech Biomed Eng, vol. 7, no. 3, pp. 159–166, 2004.
+
+The results are the forces and moments applied by the distal segment on the
+proximal articulation, in the global coordinates system.
+
+```{warning}
+This is an experimental module, its API and behaviour may change in the future. This module will be redesigned to make use of the `anthropometrics` module currently in development.
+```
+
+```{code-cell} ipython3
+import kineticstoolkit.lab as ktk
+import matplotlib.pyplot as plt
+```
+
+## Loading sample data
+
+In this tutorial, we will load propulsion data from a sprint in Wheelchair Basketball. Kinetics were recorded using an instrumented wheel (SmartWheel) and kinematics were recorded using an optoelectronic system (Optitrack). Both are synchronized.
+
+```{code-cell}
+# Load kinetics
+kinetics = ktk.load(
+    ktk.config.root_folder +
+    '/data/inversedynamics/basketball_kinetics.ktk.zip')
+
+# Change some signs to match (locally): x anterior, # y up, z lateral (right)
+kinetics.data['Forces'][:, 0] = -kinetics.data['Forces'][:, 0]
+kinetics.data['Forces'][:, 1] = -kinetics.data['Forces'][:, 1]
+
+kinetics.data['Moments'][:, 0] = -kinetics.data['Moments'][:, 0]
+kinetics.data['Moments'][:, 1] = -kinetics.data['Moments'][:, 1]
+
+# Keep everything between sync and stop recording events.
+kinetics = kinetics.get_ts_between_events('sync', 'stop')
+
+# Look at the data
+plt.subplot(1, 2, 1)
+kinetics.plot('Forces')
+plt.subplot(1, 2, 2)
+kinetics.plot('Moments')
+```
+
+```{code-cell}
+# Load kinematics
+kinematics = ktk.load(
+    ktk.config.root_folder +
+    '/data/inversedynamics/basketball_kinematics.ktk.zip')['Kinematics']
+
+# Extract markers
+markers = kinematics['Markers']
+markers.events = kinetics.events
+
+# Extract frames
+frames = kinematics['ReferenceFrames']
+frames.events = kinetics.events
+
+# Keep everything between sync and stop recording events.
+markers = markers.get_ts_between_events('sync', 'stop')
+frames = frames.get_ts_between_events('sync', 'stop')
+
+# Plot one sample
+markers.plot('RadialStyloidR')
+
+# Check what markers are available
+markers.data
+```
+
+```{code-cell}
+# Check what frames are available
+frames.data
+```
+
+## Put everything relevant into one, synchronized TimeSeries
+
+We will now create a single TimeSeries with every relevant signal. Note that for kinematics, we use `merge` instead of simply copying the data, because kinetics and kinematics have different sampling rate and therefore the data have different lengths. Merging takes care of resampling.
+
+```{code-cell}
[…]
