+++ doc/conf.py
@@ -55,6 +55,7 @@ autodoc_default_options = {
 }
 
 myst_enable_extensions = [
+    'colon_fence',
     'dollarmath',
     'html_admonition',
 ]
