+++ doc/21_timeseries/timeseries.rst
@@ -1,20 +1,0 @@
-TimeSeries
-==========
-
-The TimeSeries_ class is used in most Kinetic Toolkit's modules and is the only
-class that users have to learn after python standard types and numpy arrays.
-
-.. _TimeSeries: ../api/kineticstoolkit.TimeSeries.rst
-
-The three roles of the TimeSeries class are to:
-
-1. organize multidimensional data in time;
-2. manage events;
-3. associate metadata to data.
-
-.. toctree::
-    :maxdepth: 2
-
-    01_basics
-    02_manipulating
-    03_dataframes
