+++ doc/21_timeseries/00_timeseries.md
@@ -1,0 +1,15 @@
+# TimeSeries
+
+The [TimeSeries](../api/kineticstoolkit.TimeSeries.rst) class is used in most Kinetic Toolkit's modules and is the only class that users have to learn after python standard types and numpy arrays.
+
+The three roles of the TimeSeries class are to:
+
+1. organize multidimensional data in time;
+2. manage events;
+3. associate metadata to data.
+
+```{toctree}
+01_basics
+02_manipulating
+03_dataframes
+```
