+++ doc/25_geometry/00_geometry.md
@@ -1,0 +1,10 @@
+# Rigid body geometry
+
+The [geometry](../api/kineticstoolkit.geometry.rst) module simplifies the calculation of linear algebric operations on series of points, vectors and transformation matrices.
+
+In this tutorial, we will do a quick starter on rigid body basics (points, vectors, frames, homogeneous transforms), how to navigate between frames and local/global coordinates, and how to extract rotation angles from a series of transformation matrices.
+
+```{toctree}
+01_basics
+02_dimension_conventions
+```
