+++ doc/23_filters/filters.rst
@@ -1,14 +1,0 @@
-Filtering
-=========
-
-The filters_ module wraps some filters to use directly with Kinetics Toolkit's TimeSeries objects. These filters are convenience wrappers for numpy's and scipy's filters.
-
-.. _filters: ../api/kineticstoolkit.filters.rst
-
-.. toctree::
-    :maxdepth: 2
-
-    01_butter
-    02_smooth
-    03_savgol
-    04_median
