+++ doc/43_dev/00_tutorials.md
@@ -1,0 +1,7 @@
+# Tutorials for unreleased features
+
+```{toctree}
+01_reconstruction
+02_anthropometrics
+03_inversedynamics
+```
