+++ doc/23_filters/00_filters.md
@@ -1,0 +1,10 @@
+# Filtering
+
+The [filters](../api/kineticstoolkit.filters.rst) module wraps some filters to use directly with Kinetics Toolkit's TimeSeries objects. These filters are convenience wrappers for numpy's and scipy's filters.
+
+```{toctree}
+01_butter
+02_smooth
+03_savgol
+04_median
+```
