+++ doc/27_pushrimkinetics/00_pushrimkinetics.md
@@ -1,0 +1,196 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+# Wheelchair kinetics
+
+The [pushrimkinetics](../api/kineticstoolkit.pushrimkinetics.rst) module allows processing kinetics data from instrumented wheelchair wheels such as the SmartWheel.
+
+```{code-cell} ipython3
+import kineticstoolkit.lab as ktk
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+```
+
+## Read data from file
+
+The first step is to load data from a file, using the [pushrimkinetics.read_file()](../api/kineticstoolkit.pushrimkinetics.read_file.rst) function.
+
+```{code-cell} ipython3
+filename = (
+    ktk.config.root_folder +
+    '/data/pushrimkinetics/sample_swl_overground_propulsion_withrubber.csv')
+
+kinetics = ktk.pushrimkinetics.read_file(filename, file_format='smartwheel')
+```
+
+Let see what we loaded:
+
+```{code-cell} ipython3
+kinetics
+```
+
+```{code-cell} ipython3
+kinetics.data
+```
+
+```{code-cell} ipython3
+plt.subplot(2, 1, 1)
+kinetics.plot('Forces')
+plt.subplot(2, 1, 2)
+kinetics.plot('Moments')
+plt.tight_layout()
+```
+
+## Calculate forces and moments
+
+If the loaded data doesn't include forces and moments but only raw data (for exemple, when loading data from a SmartWheel's SD card), we must calculate the forces and moments based on a calibration matrix. The function [pushrimkinetics.calculate_forces_and_moments()](../api/kineticstoolkit.pushrimkinetics.calculate_forces_and_moments.rst) performs this calculation. In this example, we use a calibration matrix that is included in `ktk.pushrimkinetics.CALIBRATION_MATRICES`, and we express the calculated forces and moments in a reference frame that is orthogonal to the ground (levelled using the wheel's angular encoder).
+
+```{code-cell} ipython3
+calibration_matrices = ktk.pushrimkinetics.CALIBRATION_MATRICES['SmartWheel_123']
+calibration_matrices
+```
+
+```{code-cell} ipython3
+new_kinetics = ktk.pushrimkinetics.calculate_forces_and_moments(
+    kinetics,
+    gains=calibration_matrices['gains'],
+    offsets=calibration_matrices['offsets'],
+    transducer='smartwheel',
+    reference_frame='hub')
+
+plt.subplot(2, 1, 1)
+new_kinetics.plot('Forces')
+plt.subplot(2, 1, 2)
+new_kinetics.plot('Moments')
+plt.tight_layout()
+```
+
+We observe some sign differences here. In fact, the SmartWheel softwares inverts some signals based on the SmartWheel's sign convention and the side of the wheel. When the forces and moments are reconstructed from raw data, these sign changes do not happen, which explains these discrepancies.
+
+## Removing dynamic offsets in forces and moments
+
+We observe in the last graphs that dynamic (sinusoidal) offsets are presents mostly in the forces but also in the moments. We can auto-remove these offsets using [pushrimkinetics.remove_offsets()](../api/kineticstoolkit.pushrimkinetics.remove_offsets.rst).
+
+Let's apply this function on the data we just loaded.
+
+```{code-cell} ipython3
+plt.subplot(2, 1, 1)
+kinetics.plot('Forces')
+plt.title('Before removing offsets')
+plt.tight_layout()
+```
+
+```{code-cell} ipython3
[…]

[…]
