### 2021-11-20 25f668a
[doc_to_md] Added the figure template and started to put figures in {figure} directives (in module geometry tutorials)
