+++ doc/_templates/Figure.md
@@ -1,0 +1,6 @@
+:::{figure-md} optional_target
+:width: 200px
+![optional_short_name](link_to_image)
+
+Figure caption
+:::
