+++ doc/_templates/Figure.md
@@ -1,6 +1,0 @@
-:::{figure-md} optional_target
-:width: 200px
-![optional_short_name](link_to_image)
-
-Figure caption
-:::
