+++ doc/02_tutorials/06_kinematics/index.md
@@ -1,0 +1,10 @@
+# Kinematics analysis
+
+The [kinematics](../../api/kineticstoolkit.kinematics.rst) module allows opening trajectories of markers from 3D acquisition files, then process those trajectory (e.g., calculating joint angles).
+
+```{toctree}
+:glob:
+:maxdepth: 2
+
+*
+```

[…]
