+++ doc/02_tutorials/06_kinematics/01_load_visualize.md
@@ -19,7 +19,7 @@ import kineticstoolkit.lab as ktk
 
 ## Reading marker trajectories
 
-The [kinematics](../../api/kineticstoolkit.kinematics.rst) module provide the functions [kinematics.read_c3d_file()](../../api/kineticstoolkit.kinematics.read_c3d_file.rst) and [kinematics.read_n3d_file()](../../api/kineticstoolkit.kinematics.read_n3d_file.rst) to read files generated by common motion capture devices. Reading c3d files is provided by the external [ezc3d](https://github.com/pyomeca/ezc3d) package; see the "Installing" section for information on how to install it.
+The [kinematics](/api/kineticstoolkit.kinematics.rst) module provide the functions [kinematics.read_c3d_file()](/api/kineticstoolkit.kinematics.read_c3d_file.rst) and [kinematics.read_n3d_file()](/api/kineticstoolkit.kinematics.read_n3d_file.rst) to read files generated by common motion capture devices. Reading c3d files is provided by the external [ezc3d](https://github.com/pyomeca/ezc3d) package; see the "Installing" section for information on how to install it.
 
 Let's read the c3d file we recorded in this acquisition:
 
@@ -38,7 +38,7 @@ markers.data
 
 ## Visualizing the markers
 
-Now we'll take a look at this acquisition using [ktk.Player](../../api/kineticstoolkit.Player.rst), a matplotlib-based interactive user interface aimed at visualizing markers, rigid bodies and segments in three dimensions.
+Now we'll take a look at this acquisition using [ktk.Player](/api/kineticstoolkit.Player.rst), a matplotlib-based interactive user interface aimed at visualizing markers, rigid bodies and segments in three dimensions.
 
 For the `Player` class to be interactive, you must select an interactive backend for IPython. See the "Installing" section for more information. It is also possible to set a point of view and generate a fixed animation in inline matplotlib environments such as Jupyter notebooks. This is what we will do in these tutorials.
 
