+++ doc/02_tutorials/01_timeseries/01_basics.md
@@ -38,7 +38,7 @@ ts
 ts.data
 ```
 
-TimeSeries can be [plotted](../../api/kineticstoolkit.TimeSeries.plot.rst) directly using Matplotlib:
+TimeSeries can be [plotted](/api/kineticstoolkit.TimeSeries.plot.rst) directly using Matplotlib:
 
 ```{code-cell}
 ts.plot()
@@ -102,7 +102,7 @@ The `time_info` property associates metadata to the time vector. It is a diction
 ts.time_info
 ```
 
-Similarly, the `data_info` property associates metadata to data. This property is a dictionary of dictionaries, where the outer key corresponds to the data key, and the inner key is the metadata. The [TimeSeries.add_data_info()](../../api/kineticstoolkit.TimeSeries.add_data_info.rst) method eases the management of `data_info`.
+Similarly, the `data_info` property associates metadata to data. This property is a dictionary of dictionaries, where the outer key corresponds to the data key, and the inner key is the metadata. The [TimeSeries.add_data_info()](/api/kineticstoolkit.TimeSeries.add_data_info.rst) method eases the management of `data_info`.
 
 ```{code-cell}
 ts = ts.add_data_info('Forces', 'Unit', 'N')
@@ -115,7 +115,7 @@ ts.data_info
 ts.data_info['Forces']
 ```
 
-Unless explicitly mentioned, metadata is not used for calculation and is optional. It is simply a way to clarify the data by adding information to it. Some functions however read metadata: for example, the [TimeSeries.plot()](../../api/kineticstoolkit.TimeSeries.plot.rst) method looks for possible `Unit` metadata and prints it on the y axis.
+Unless explicitly mentioned, metadata is not used for calculation and is optional. It is simply a way to clarify the data by adding information to it. Some functions however read metadata: for example, the [TimeSeries.plot()](/api/kineticstoolkit.TimeSeries.plot.rst) method looks for possible `Unit` metadata and prints it on the y axis.
 
 ```{code-cell}
 ts.plot()
@@ -126,9 +126,9 @@ ts.plot()
 In the figure above, we see that the TimeSeries contains cyclic data that could be characterized by events. A first spike was generated at about 4 seconds: this event corresponds to a synchronization signal that we generated by gently impacting the instrumented pushrim. Thereafter, we see a series of pushes and recoveries. We will add these events to the TimeSeries.
 
 There are several ways to edit the events of a TimeSeries:
-- Editing events manually, using the [TimeSeries.add_event()](../../api/kineticstoolkit.TimeSeries.add_event.rst) and [TimeSeries.remove_event()](../../api/kineticstoolkit.TimeSeries.remove_event.rst) methods;
-- Editing events interactively, using the [TimeSeries.ui_edit_events()](../../api/kineticstoolkit.TimeSeries.ui_edit_events.rst) method;
-- Adding events automatically, for example using the [cycles](../24_cycles/00_cycles.rst) module that can detect cycles automatically.
+- Editing events manually, using the [TimeSeries.add_event()](/api/kineticstoolkit.TimeSeries.add_event.rst) and [TimeSeries.remove_event()](/api/kineticstoolkit.TimeSeries.remove_event.rst) methods;
+- Editing events interactively, using the [TimeSeries.ui_edit_events()](/api/kineticstoolkit.TimeSeries.ui_edit_events.rst) method;
+- Adding events automatically, for example using the [cycles](/api/kineticstoolkit.cycles.rst) module that can detect cycles automatically.
 
 In this tutorial, we will add the events manually.
 
@@ -164,7 +164,7 @@ ts.plot()
 
 ### Using events to synchronize TimeSeries
 
[…]
