+++ doc/02_tutorials/06_kinematics/index.md
@@ -1,6 +1,6 @@
 # Kinematics analysis
 
-The [kinematics](../../api/kineticstoolkit.kinematics.rst) module allows opening trajectories of markers from 3D acquisition files, then process those trajectory (e.g., calculating joint angles).
+The [kinematics](/api/kineticstoolkit.kinematics.rst) module allows opening trajectories of markers from 3D acquisition files, then process those trajectory (e.g., calculating joint angles).
 
 ```{toctree}
 :glob:
