+++ doc/02_tutorials/03_filters/01_butter.md
@@ -11,7 +11,7 @@ kernelspec:
 
 # Removing frequencies using a Butterworth filter
 
-The Butterworth filter may be the most used filter in biomechanics. It targets ranges of frequencies to remove from the signal's frequency spectrum. A classic use is to estimate the frequency range of both the data and noise, then use the filter to keep most of the data's frequency range while filtering out most of the noise's frequency range. In this tutorial, we will see how to apply Butterworth filters on TimeSeries data, using the [filters.butter()](../../api/kineticstoolkit.filters.butter.rst) function.
+The Butterworth filter may be the most used filter in biomechanics. It targets ranges of frequencies to remove from the signal's frequency spectrum. A classic use is to estimate the frequency range of both the data and noise, then use the filter to keep most of the data's frequency range while filtering out most of the noise's frequency range. In this tutorial, we will see how to apply Butterworth filters on TimeSeries data, using the [filters.butter()](/api/kineticstoolkit.filters.butter.rst) function.
 
 ```{code-cell}
 import kineticstoolkit.lab as ktk
