### 2021-11-23 6bd223a
[master] Removed doc folder and dev.build_website() since website construction is moved to the kineticstoolkit_doc repository
