+++ kineticstoolkit/dev.py
@@ -130,56 +130,6 @@ def run_doc_tests() -> None:  # pragma: no cover
     os.chdir(cwd)
 
 
-def build_website(clean: bool = False) -> None:  # pragma: no cover
-    """
-    Build the website using sphinx.
-
-    Set clean to True to `make clean` beforehand.
-
-    """
-    cwd = os.getcwd()
-    os.chdir(kineticstoolkit.config.root_folder + '/doc')
-
-    if clean:
-        shutil.rmtree(kineticstoolkit.config.root_folder + '/doc/api',
-                      ignore_errors=True)
-        subprocess.call(['make', 'clean'],
-                        env=kineticstoolkit.config.env)
-
-    # Generate site
-    print('Generating site...')
-    subprocess.call(['make', 'html'], env=kineticstoolkit.config.env)
-
-    # Move site to documentation repository
-    if kineticstoolkit.config.version == 'master':
-        doc_folder = (
-            kineticstoolkit.config.root_folder
-            + '/../kineticstoolkit_doc/master'
-        )
-    else:
-        doc_folder = (
-            kineticstoolkit.config.root_folder
-            + '/../kineticstoolkit_doc/stable'
-        )
-
-    try:
-        shutil.rmtree(doc_folder)
-    except Exception:
-        pass
-
-    os.makedirs(doc_folder, exist_ok=True)
-    os.rename(
-        kineticstoolkit.config.root_folder + '/doc/_build/html',
-        doc_folder,
-    )
-
-    # Open site
-    webbrowser.open_new_tab(
-        'file://' + doc_folder + '/index.html'
-    )
-    os.chdir(cwd)
-
-
 def compile_for_pypi() -> None:  # pragma: no cover
     """Compile for PyPi."""
     shutil.rmtree(kineticstoolkit.config.root_folder + '/dist',
