+++ doc/02_tutorials/06_kinematics/02_joint_angles.md
@@ -1,248 +1,0 @@
----
-jupytext:
-  text_representation:
-    extension: .md
-    format_name: myst
-kernelspec:
-  display_name: Python 3 (ipykernel)
-  language: python
-  name: python3
----
-
-# Calculating joint angles from a kinematic acquisition
-
-In this tutorial, we will calculate the elbow joint angles in the acquisition presented in the previous section. At the end of the tutorial, we will have plotted curves of elbow flexion and pronation angles during the complete acquisition.
-
-Such an analysis is composed of the following steps:
-
-**Step 1.** Read the marker trajectories
-
-For each of the analyzed joint:
-
-**Step 2.** We calculate series of frames to express the orientation of the two segments that compose this joint (how are both segments oriented in space);
-
-**Step 3.** We express a series of homogeneous transforms between the proximal and distal frames (how the distal segment is oriented in space compared to the proximal segment);
-
-**Step 4.** We extract the series of Euler angles from these homogeneous transforms (what series of three angles allow such relative orientation);
-
-**Step 5.** We interpret these Euler angle as rotations around anatomical axes.
-
-```{code-cell}
-import kineticstoolkit.lab as ktk
-```
-
-## Read and visualize marker trajectories
-
-We proceed exactly as in the previous tutorial:
-
-```{code-cell} ipython3
-# Read the markers
-markers = ktk.kinematics.read_c3d_file(
-    ktk.config.root_folder + '/data/kinematics/racing_full.c3d')
-
-# Set the point of view for 3D visualization
-viewing_options = {
-    'zoom': 3.5,
-    'azimuth': 0.8,
-    'elevation': 0.16,
-    'translation': (0.2, -0.7)
-}
-
-# Define the segments for easier visualization
-segments = {
-    'ArmR': {
-        'Color': [1, 0.25, 0],
-        'Links': [['AcromionR', 'MedialEpicondyleR'],
-                  ['AcromionR', 'LateralEpicondyleR'],
-                  ['AcromionR', 'OlecraneR']]
-    },
-    'ForearmR': {
-        'Color': [1, 0.5, 0],
-        'Links': [['MedialEpicondyleR', 'RadialStyloidR'],
-                  ['MedialEpicondyleR', 'UlnarStyloidR'],
-                  ['LateralEpicondyleR', 'RadialStyloidR'],
-                  ['LateralEpicondyleR', 'UlnarStyloidR'],
-                  ['OlecraneR', 'RadialStyloidR'],
-                  ['OlecraneR', 'UlnarStyloidR'],
-                  ['UlnarStyloidR', 'RadialStyloidR']]
-    }
-}
-
-# Create the player
-player = ktk.Player(markers, segments=segments, **viewing_options)
-
-# Show one second (only needed in Notebooks)
-player.to_html5(start_time=0, stop_time=1)
-```
-
-## Create local coordinate systems
-
-To calculate the elbow angles, we need to express frames for the arm and forearm segments. The elbow angles will be extracted from the homogeneous transform between both frames.
-
-We will use the recommendations of the International Society of Biomechanics [1] to define these local coordinate systems:
-
-### Arm coordinate system
-
-Following the ISB recommendations, the local coordinate system for the humerus is defined as follow:
-
-**1. The origin is GH (glenohumeral joint).**
-
-We don't have a marker for the glenohumeral join, but we have one for the acromion. We will therefore approximate GH by the acromion.
-
-```{code-cell}
-origin = markers.data['AcromionR']
-```
-
-**2. The y axis is the line between GH and the midpoint of EL (lateral elbow epicondyle) and EM (medial elbow epicondyle), pointing to GH**
-
-```{code-cell}
-y = (markers.data['AcromionR'] -
[…]
