+++ doc/02_tutorials/01_timeseries/index.md
@@ -1,16 +1,0 @@
-# TimeSeries
-
-The [TimeSeries](/api/kineticstoolkit.TimeSeries.rst) class is used in most Kinetic Toolkit's modules and is the only class that users have to learn after python standard types and numpy arrays.
-
-The three roles of the TimeSeries class are to:
-
-1. organize multidimensional data in time;
-2. manage events;
-3. associate metadata to data.
-
-```{toctree}
-:glob:
-:maxdepth: 2
-
-*
-```

[…]
