### 2021-12-21 1015ce8
#31 Added some basic tests in TimeSeriesEvent
