### 2021-12-21 c1986ef
#31 Added some basic tests in TimeSeriesEvent
