+++ kineticstoolkit/pushrimkinetics.py
@@ -313,12 +313,12 @@ def calculate_forces_and_moments(
     For standard force cells (with each channel being a raw value
     corresponding to Fx, Fy, Fz, Mx, My, Mz, respectively), calculates
     the forces and moments using a sensitivity matrix (gains) and an
-    offset vector (offsets):
-
-    ``[Fx, Fy, Fz, Mx, My, Mz] = gains @ channels + offsets``
+    offset vector (offsets).
 
     For SmartWheel, calculates the forces and moments using a gain
     vector (gains) and an offset vector (offsets).
+
+        [Fx, Fy, Fz, Mx, My, Mz] = gains @ channels + offsets
 
     Parameters
     ----------
@@ -342,18 +342,6 @@ def calculate_forces_and_moments(
     TimeSeries
         A copy of the input TimeSeries, with the added 'Forces'
         and 'Moments' data keys.
-
-    Note
-    ----
-    Some calibration matrices are provided as examples in the
-    ``pushrimkinetics.CALIBRATION_MATRICES`` dictionary. This dictionary can
-    be used directly using dict unpacking. For example::
-
-        ktk.pushrimkinetics.calculate_force_and_moments(
-            kinetics,
-            **ktk.pushrimkinetics.CALIBRATION_MATRICES['SmartWheel_123'],
-            reference_frame='level'
-        )
 
     """
     # Calculate the forces and moments and add to the output
@@ -502,143 +490,60 @@ def calculate_power(tsin: TimeSeries, /) -> TimeSeries:
     return tsout
 
 
-#--- Deprecated functions ---#
-@dead(since="October 2020",
-      until="December 2021",
-      details="Please use ktk.cycles.detect_cycles instead.")
-def detect_pushes(
-        tsin: TimeSeries, /, *,
-        push_threshold: float = 5.0,
-        recovery_threshold: float = 2.0,
-        min_push_time: float = 0.1,
-        min_push_force: float = 30.0) -> TimeSeries:
-    """
-    Detect pushes and recoveries automatically.
+# Deprecated constants:
 
-    Parameters
-    ----------
-    tsin
-        Input TimeSeries that must contain a 'Forces' key in its data dict.
-    push_threshold
-        Optional. The total force over which a push phase is triggered, in
-        newton.
-    recovery_threshold
-        Optional. The total force under which a recovery phase is triggered,
-        in newton.
-    min_push_time
-        Optional. The minimum time required for a push time, in seconds.
-        Detected pushes that last less than this minimum time are removed from
-        the push analysis.
-    min_recovery_time
-        Optional. The minimum time required for a recovery time, in seconds.
-        Detected recoveries that last less than this minimum time are removed
-        from the push analysis.
-    min_push_force
-        Optional. The minimum total push force in N under which the detected
-        push is discarded. For example, if the user puts their hands on the
-        pushrim before starting propelling, this may be detected as a push.
[…]
