### 2022-01-13 a562fa0
[master] Performance enhancement. TimeSeries: 2x speed boost on all get_ts_[...] methods. Important speed boost on ktk.cycles.detect_cycles().
