### 2022-02-16 0e2de73
[master] ktk.doc.download_sample_data is not ktk.doc.download, and it's not used everywhere in the documentation.
