### 2022-02-17 222495e
[master] #74: Simplified data files available for downloading
