+++ tests/test_doc.py
@@ -29,10 +29,12 @@ import kineticstoolkit as ktk
 def test_download_sample_data():
     """Test ktk.doc.download_sample_data."""
     # Download the usual way
-    file1 = ktk.doc.download('noisy_signals.ktk.zip')
+    file1 = ktk.doc.download(
+        'filters_noisy_signals.ktk.zip'
+    )
     # Use the local version
     file2 = ktk.doc.download(
-        'noisy_signals.ktk.zip',
+        'filters_noisy_signals.ktk.zip',
         force_download=True
     )
     assert file1 != file2
