### 2022-02-17 a64b5ec
[master] #74: Simplified data files available for downloading
