### 2022-02-22 673f104
[master] Closes #81 Developer issue: switch from master to stable on the fly
