+++ kineticstoolkit/dev.py
@@ -38,38 +38,6 @@ import subprocess
 import shutil
 import webbrowser
 import doctest
-import multiprocessing
-import time
-import json
-
-
-def update_version() -> None:  # pragma: no cover
-    """Update VERSION based on the current branch name."""
-    # Get the current branch name
-    cwd = os.getcwd()
-    os.chdir(kineticstoolkit.config.root_folder)
-    branch_name = subprocess.check_output(
-        ['git', 'branch', '--show-current']).decode()
-    if 'stable' in branch_name:
-        shutil.copy(
-            kineticstoolkit.config.root_folder
-            + '/kineticstoolkit/STABLE_VERSION',
-            kineticstoolkit.config.root_folder
-            + '/kineticstoolkit/VERSION')
-    else:
-        shutil.copy(
-            kineticstoolkit.config.root_folder
-            + '/kineticstoolkit/DEVELOPMENT_VERSION',
-            kineticstoolkit.config.root_folder
-            + '/kineticstoolkit/VERSION')
-
-    # Update version in config module
-    with open(
-            kineticstoolkit.config.root_folder
-            + '/kineticstoolkit/VERSION', 'r') as fid:
-        kineticstoolkit.config.version = fid.read()
-
-    os.chdir(cwd)
 
 
 def run_unit_tests() -> None:  # pragma: no cover
@@ -154,7 +122,6 @@ def upload_to_pypi() -> None:  # pragma: no cover
 
 def run_tests() -> None:  # pragma: no cover
     """Run all testing and building functions."""
-    update_version()
     run_style_formatter()
     run_doc_tests()
     run_static_type_checker()
