### 2022-02-22 9a73c89
[master] #53 Player: segments is not interconnections.
