+++ kineticstoolkit/player.py
@@ -59,9 +59,9 @@ class Player:
         or a rigid body pose expressed as a Nx4x4 array. Multiple TimeSeries
         can be provided, e.g., ktk.Player(markers, rigid_bodies)
 
-    segments
-        Optional. Used to draw lines between markers. Each key corresponds to a
-        segment, where a segment is another dict with the following keys:
+    interconnections
+        Optional. Each key corresponds to an inerconnection between markers,
+        where one interconnection is another dict with the following keys:
 
         - Links: list of list of 2 str, where each str is a marker
           name. For example, to link Marker1 to Marker2 and
@@ -108,19 +108,26 @@ class Player:
 
     def __init__(self,
                  *ts: TimeSeries,
-                 segments: Dict[str, Dict[str, Any]] = {},
+                 interconnections: Dict[str, Dict[str, Any]] = {},
                  current_frame: int = 0,
                  marker_radius: float = 0.008,
                  axis_length: float = 0.1,
                  axis_width: float = 3.0,
-                 segment_width: float = 1.5,
+                 interconnection_width: float = 1.5,
                  zoom: float = 1.0,
                  azimuth: float = 0.0,
                  elevation: float = 0.2,
                  translation: Union[Sequence[float], np.ndarray] = (0.0, 0.0),
                  target: Union[Sequence[float], np.ndarray] = (0.0, 0.0, 0.0),
                  track: bool = False,
-                 perspective: bool = True):
+                 perspective: bool = True,
+                 **kwargs):
+
+        # Allow the older `segments` and interconnection_width keywords:
+        if 'segments' in kwargs and interconnections == {}:
+            interconnections = kwargs['segments']
+        if 'segment_width' in kwargs:
+            interconnection_width = kwargs['segment_width']
 
         # ---------------------------------------------------------------
         # Set self.n_frames and self.time, and verify that we have at least
@@ -167,8 +174,8 @@ class Player:
             np.eye(4, 4)[np.newaxis, :, :], self.n_frames, axis=0)
 
         # ---------------------------------------------------------------
-        # Assign the segments
-        self.segments = segments
+        # Assign the interconnections
+        self.interconnections = interconnections
 
         # ---------------------------------------------------------------
         # Other initalizations
@@ -176,7 +183,7 @@ class Player:
         self.marker_radius = marker_radius
         self.axis_length = axis_length
         self.axis_width = axis_width
-        self.segment_width = segment_width
+        self.interconnection_width = interconnection_width
         self.zoom = zoom
         self.azimuth = azimuth
         self.elevation = elevation
@@ -197,7 +204,7 @@ class Player:
         self.objects['PlotRigidBodiesY'] = None
         self.objects['PlotRigidBodiesZ'] = None
         self.objects['PlotGroundPlane'] = None
-        self.objects['PlotSegments'] = dict()
+        self.objects['PlotInterconnections'] = dict()
 
         self.objects['Figure'] = None
         self.objects['Axes'] = None
@@ -240,7 +247,7 @@ class Player:
             '''
 
         self._create_figure()
-        self._create_segments()
+        self._create_interconnections()
         self._create_markers()
         self._create_ground_plane()
         self._first_refresh()
@@ -294,15 +301,15 @@ class Player:
         self.objects['Figure'].canvas.mpl_connect(
             'motion_notify_event', self._on_mouse_motion)
 
-    def _create_segments(self) -> None:
-        """Create the segments plots in the player's figure."""
-        if self.segments is not None:
-            for segment in self.segments:
-                self.objects['PlotSegments'][segment] = \
+    def _create_interconnections(self) -> None:
+        """Create the interconnections plots in the player's figure."""
+        if self.interconnections is not None:
+            for interconnection in self.interconnections:
[…]
