+++ kineticstoolkit/player.py
@@ -72,8 +72,8 @@ class Player:
           link. Color must be a valid value for matplotlib's
           plots.
 
-    current_frame
-        Optional. Sets the inital frame number to show.
+    current_index
+        Optional. Sets the inital index number to show.
 
     marker_radius
         Optional. Sets the marker radius as defined by matplotlib.
@@ -98,7 +98,7 @@ class Player:
 
     track
         Optional. False to keep the scene static, True to track the last
-        selected marker when changing frame.
+        selected marker when changing index.
 
     perspective
         Optional. True to draw the scene using perspective, False to draw the
@@ -109,7 +109,7 @@ class Player:
     def __init__(self,
                  *ts: TimeSeries,
                  interconnections: Dict[str, Dict[str, Any]] = {},
-                 current_frame: int = 0,
+                 current_index: int = 0,
                  marker_radius: float = 0.008,
                  axis_length: float = 0.1,
                  axis_width: float = 3.0,
@@ -123,17 +123,19 @@ class Player:
                  perspective: bool = True,
                  **kwargs):
 
-        # Allow the older `segments` and interconnection_width keywords:
+        # Allow older variable names
         if 'segments' in kwargs and interconnections == {}:
             interconnections = kwargs['segments']
         if 'segment_width' in kwargs:
             interconnection_width = kwargs['segment_width']
+        if 'current_frame' in kwargs:
+            current_index = kwargs['current_frame']
 
         # ---------------------------------------------------------------
-        # Set self.n_frames and self.time, and verify that we have at least
+        # Set self.n_indexes and self.time, and verify that we have at least
         # markers or rigid bodies.
         self.time = ts[0].time
-        self.n_frames = len(ts[0].time)
+        self.n_indexes = len(ts[0].time)
 
         # ---------------------------------------------------------------
         # Assign the markers and rigid bodies
@@ -171,7 +173,7 @@ class Player:
 
         # Add the origin to the rigid bodies
         self.rigid_bodies.data['Global'] = np.repeat(
-            np.eye(4, 4)[np.newaxis, :, :], self.n_frames, axis=0)
+            np.eye(4, 4)[np.newaxis, :, :], self.n_indexes, axis=0)
 
         # ---------------------------------------------------------------
         # Assign the interconnections
@@ -179,7 +181,7 @@ class Player:
 
         # ---------------------------------------------------------------
         # Other initalizations
-        self.current_frame = current_frame
+        self.current_index = current_index
         self.marker_radius = marker_radius
         self.axis_length = axis_length
         self.axis_width = axis_width
@@ -229,8 +231,8 @@ class Player:
             ----------------------------------------------------
             KEYBOARD COMMANDS
             show/hide this help : h
-            previous frame      : left
-            next frame          : right
+            previous index      : left
+            next index          : right
             previous second     : shift+left
             next second         : shift+right
             play/pause          : space
@@ -426,7 +428,7 @@ class Player:
         return rotated_points_3d[:, 0:2]
 
     def _update_markers_and_interconnections(self) -> None:
-        # Get a Nx4 matrices of every marker at the current frame
+        # Get a Nx4 matrices of every marker at the current index
         markers = self.markers
         if markers is None:
             return
@@ -453,7 +455,7 @@ class Player:
                 else:
                     color = 'w'
 
-                these_coordinates = markers.data[marker][self.current_frame]
+                these_coordinates = markers.data[marker][self.current_index]
                 markers_data[color][i_marker] = these_coordinates
                 interconnection_markers[marker] = these_coordinates
 
@@ -508,7 +510,7 @@ class Player:
[…]
